Git Gud v0.1.0 GitGud.Authorization

Conveniences for authorization and resource loading.

Functions

authorized?(user, resource, action)

authorized?(GitGud.User.t() | nil, GitGud.AuthorizationPolicies.t(), atom()) ::
  boolean()

Returns true if user is allowed to perform action on resource; otherwhise returns false.

enforce_policy!(user, resource, action, default \\ nil)

enforce_policy!(
  GitGud.User.t() | nil,
  GitGud.AuthorizationPolicies.t(),
  atom(),
  any()
) :: GitGud.AuthorizationPolicies.t() | any()

Same as enforce_policy/3 but returns the resource or default if the policy cannot be enforced.

enforce_policy(user, resource, action)

enforce_policy(GitGud.User.t() | nil, GitGud.AuthorizationPolicies.t(), atom()) ::
  {:ok, GitGud.AuthorizationPolicies.t()} | {:error, :unauthorized}

Enforces the authorization policy.

filter(user, resources, action)

filter(GitGud.User.t() | nil, [GitGud.AuthorizationPolicies.t()], atom()) :: [
  GitGud.AuthorizationPolicies.t()
]

Filters the given list of resources, i.e. returns only those for which enforce_policy/3 applies.