GitGud.Email (git.limo v0.3.8)

Email schema and helper functions.

A GitGud.Email is used for different tasks such as user authentication & verification, notifications, identification of Git commit authors & commiters, etc.

Types

t()

Specs

t() :: %GitGud.Email{
  __meta__: term(),
  address: binary(),
  id: pos_integer(),
  inserted_at: NaiveDateTime.t(),
  user: GitGud.User.t(),
  user_id: pos_integer(),
  verified: boolean(),
  verified_at: NaiveDateTime.t()
}

Functions

changeset(email, params \\ %{})

Specs

changeset(t(), map()) :: Ecto.Changeset.t()

Returns a changeset for the given params.

create(user, params)

Specs

create(GitGud.User.t(), map() | keyword()) ::
  {:ok, t()} | {:error, Ecto.Changeset.t()}

Creates a new email with the given params.

{:ok, email} = GitGud.Email.create(user, address: "m.flach@almightycouch.com")

This function validates the given params using changeset/2.

create!(user, params)

Specs

create!(GitGud.User.t(), map() | keyword()) :: t()

Similar to create/1, but raises an Ecto.InvalidChangesetError if an error occurs.

delete(email)

Specs

delete(t()) :: {:ok, t()} | {:error, Ecto.Changeset.t()}

Deletes the given email.

{:ok, email} = GitGud.Email.delete(email)

delete!(email)

Specs

delete!(t()) :: t()

Similar to delete!/1, but raises an Ecto.InvalidChangesetError if an error occurs.

verification_changeset(email)

Specs

verification_changeset(t()) :: Ecto.Changeset.t()

Returns a changeset for verifying the given email.

verify(email)

Specs

verify(t()) :: {:ok, t()} | {:error, Ecto.Changeset.t()}

Verifies the given email.

{:ok, email} = GitGud.Email.verify(email)

This function verifies the given email and ensures that the associated user is updated accordingly.

verify!(email)

Specs

verify!(t()) :: t()

Similar to verify/1, but raises an Ecto.InvalidChangesetError if an error occurs.