Git Gud v0.1.0 GitGud.GitDiff

Defines a Git diff.

Types

t()

t() :: %GitGud.GitDiff{__git__: GitRekt.Git.diff(), repo: GitGud.Repo.t()}

Functions

delta_count(diff)

delta_count(t()) :: {:ok, non_neg_integer()} | {:error, term()}

Returns the number of deltas in the given diff.

deltas(diff)

deltas(t()) ::
  {:ok,
   [
     {GitRekt.Git.diff_delta(),
      [{GitRekt.Git.diff_hunk(), [GitRekt.Git.diff_line()]}]}
   ]}
  | {:error, term()}

Returns a list of deltas for the given diff.

format(diff, diff_format \\ :patch)

format(t(), GitRekt.Git.diff_format()) :: {:ok, binary()} | {:error, term()}

Returns a binary represention of the given diff.

stats(diff)

stats(t()) ::
  {:ok, {non_neg_integer(), non_neg_integer(), non_neg_integer()}}
  | {:error, term()}

Returns stats of the given diff.