Git Gud v0.1.0 GitGud.GitTree

Defines a Git tree object.

Types

t()

t() :: %GitGud.GitTree{
  __git__: GitRekt.Git.tree(),
  oid: GitRekt.Git.oid(),
  repo: GitGud.Repo.t()
}

Functions

by_id(git_tree, oid)

by_id(t(), GitRekt.Git.oid()) ::
  {:ok, GitGud.GitTreeEntry.t()} | {:error, term()}

Returns the tree entry of the given tree, given its oid.

by_path(git_tree, path)

by_path(t(), Path.t()) :: {:ok, GitGud.GitTreeEntry.t()} | {:error, term()}

Returns the tree entry of the given tree, given its path.

entries(tree)

entries(t()) :: {:ok, Stream.t()} | {:error, term()}

Returns all the tree entries of the given tree.

pathspec_match?(git_tree, pathspec)

pathspec_match?(t(), [binary()]) :: {:ok, boolean()} | {:error, term()}

Returns true if tree matches the given pathspec; otherwise returns false.