Git Gud v0.1.0 GitGud.GraphQL.Resolvers

Module providing resolution functions for GraphQL related queries.

Functions

git_actor_type(actor, info)

git_actor_type(GitGud.User.t() | map(), Absinthe.Resolution.t()) :: atom()

Resolves the type for a given Git actor.

git_blob_size(blob, args, info)

git_blob_size(GitGud.GitBlob.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, integer()} | {:error, term()}

Resolves the content length for a given Git blob object.

git_commit_author(commit, args, info)

git_commit_author(GitGud.GitCommit.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, GitGud.User.t() | map()} | {:error, term()}

Resolves the author for a given Git commit object.

git_commit_message(commit, args, info)

git_commit_message(GitGud.GitCommit.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, binary()} | {:error, term()}

Resolves the message for a given Git commit object.

git_commit_timestamp(commit, args, info)

git_commit_timestamp(GitGud.GitCommit.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, DateTime.t()} | {:error, term()}

Resolves the timestamp for a given Git commit object.

git_history(args, source)

git_history(map(), Absinthe.Resolution.t()) ::
  {:ok, Absinthe.Relay.Connection.t()} | {:error, term()}

Resolves the commit history starting from the given Git revision object.

git_object_type(object, info)

git_object_type(GitGud.Repo.git_object(), Absinthe.Resolution.t()) :: atom()

Resolves the type for a given Git object.

git_reference_target(reference, args, info)

git_reference_target(GitGud.GitReference.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, GitGud.Repo.git_object()} | {:error, term()}

Resolves the Git target for the given Git reference object.

git_reference_type(reference, args, info)

git_reference_type(GitGud.GitReference.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, GitGud.GitReference.type()} | {:error, term()}

Resolves the type for the given Git reference object.

git_tag_author(tag, args, info)

git_tag_author(GitGud.GitTag.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, GitGud.User.t() | map()} | {:error, term()}

Resolves the author for a given Git tag object.

git_tag_message(tag, args, info)

git_tag_message(GitGud.GitTag.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, binary()} | {:error, term()}

Resolves the message for a given Git tag object.

git_tag_target(tag, args, info)

git_tag_target(GitGud.GitTag.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, GitGud.Repo.git_object()} | {:error, term()}

Resolves the Git target for the given Git tag object.

git_tag_type(tag, info)

git_tag_type(
  GitGud.GitReference.t() | GitGud.GitTag.t(),
  Absinthe.Resolution.t()
) :: {:ok, atom()} | {:error, term()}

Resolves the type for a given Git tag.

git_tree(revision, args, info)

git_tree(GitGud.Repo.git_revision(), %{}, Absinthe.Resolution.t()) ::
  {:ok, GitGud.GitTree.t()} | {:error, term()}

Resolves the tree for a given Git commit object.

git_tree_entries(args, source)

git_tree_entries(map(), Absinthe.Resolution.t()) ::
  {:ok, Absinthe.Relay.Connection.t()} | {:error, term()}

Resolves the tree entries for a given Git tree object.

git_tree_entry_target(tree_entry, args, info)

git_tree_entry_target(GitGud.Repo.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, GitGud.GitTree.t() | GitGud.GitBlob.t()} | {:error, term()}

Returns the underlying Git object for a given Git tree_entry object.

node(node_type, info)

node(map(), Absinthe.Resolution.t()) :: {:ok, map()} | {:error, term()}

Resolves a node object.

node_type(node, info)

node_type(map(), Absinthe.Resolution.t()) :: atom() | nil

Resolves a node object type.

repo_head(repo, args, info)

repo_head(GitGud.Repo.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, GitGud.GitReference.t()} | {:error, term()}

Resolves the default branch object for a given repo.

repo_owner(repo, args, info)

repo_owner(GitGud.Repo.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, GitGud.User.t()} | {:error, term()}

Resolves the owner for a given repo.

repo_ref(repo, args, info)

repo_ref(GitGud.Repo.t(), %{name: binary()}, Absinthe.Resolution.t()) ::
  {:ok, GitGud.GitReference.t()} | {:error, term()}

Resolves a Git reference object by name for a given repo.

repo_refs(args, source)

repo_refs(map(), Absinthe.Resolution.t()) ::
  {:ok, Absinthe.Relay.Connection.t()} | {:error, term()}

Resolves all Git reference objects for a given repo.

repo_tag(repo, args, info)

repo_tag(GitGud.Repo.t(), %{name: binary()}, Absinthe.Resolution.t()) ::
  {:ok, GitGud.GitReference.t() | GitGud.GitTag.t()} | {:error, term()}

Resolves a Git tag object by name for a given repo.

repo_tags(args, source)

repo_tags(map(), Absinthe.Resolution.t()) ::
  {:ok, Absinthe.Relay.Connection.t()} | {:error, term()}

Resolves all Git tag objects for a given repo.

search(args, info)

search(map(), Absinthe.Resolution.t()) ::
  {:ok, Absinthe.Relay.Connection.t()} | {:error, term()}

Resolves a list of users for a given search term.

search_result_type(arg1, info)

search_result_type(GitGud.User.t() | GitGud.Repo.t(), Absinthe.Resolution.t()) ::
  atom()

url(resource, args, info)

url(map(), %{}, Absinthe.Resolution.t()) :: {:ok, binary()} | {:error, term()}

Resolves the URL of the given resource.

user(root, args, info)

user(%{}, %{login: binary()}, Absinthe.Resolution.t()) ::
  {:ok, GitGud.User.t()} | {:error, term()}

Resolves an user object by login.

user_public_email(user, args, info)

user_public_email(GitGud.User.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, GitGud.GitReference.t()} | {:error, term()}

Resolves the public email for a given user.

user_repo(user, args, info)

user_repo(GitGud.User.t(), %{name: binary()}, Absinthe.Resolution.t()) ::
  {:ok, GitGud.Repo.t()} | {:error, term()}

Resolves a repository object by name for a given user.

user_repos(args, info)

user_repos(map(), Absinthe.Resolution.t()) ::
  {:ok, Absinthe.Relay.Connection.t()} | {:error, term()}

Resolves all repositories for a given user.