GitGud.GraphQL.Schema (git.limo v0.3.8)

GraphQL schema definition.

Functions

from_relay_id(global_id)

Specs

from_relay_id(Absinthe.Relay.Node.global_id()) :: pos_integer() | nil

Returns the source id for the given Relay global_id.

from_relay_id(global_id, info, opts \\ [])

Returns the source struct for the given Relay global_id.

run(document, opts \\ [])

Specs

run(binary() | term(), Absinthe.run_opts()) :: Absinthe.run_result()

Evaluates a query document.

to_relay_id(node)

Specs

to_relay_id(struct()) :: Absinthe.Relay.Node.global_id() | nil

Returns the Relay global id for the given node.

to_relay_id(node_type, source_id)

Specs

to_relay_id(atom() | binary(), pos_integer()) ::
  Absinthe.Relay.Node.global_id() | nil

Returns the Relay global id for the given source_id.