GitGud.IssueQuery (git.limo v0.3.8)

Conveniences for issue related queries.

Functions

by_id(id, opts \\ [])

Specs

by_id(pos_integer(), keyword()) :: GitGud.Issue.t() | nil

Returns a repository issue for the given id.

count_repo_issues(repo, opts \\ [])

Specs

count_repo_issues(GitGud.Repo.t() | pos_integer(), keyword()) ::
  non_neg_integer()

count_repo_issues([GitGud.Repo.t() | pos_integer()], keyword()) :: %{
  required(pos_integer()) => non_neg_integer()
}

Returns the number of issues for the given repo.

count_repo_issues_by_status(repo, opts \\ [])

Specs

count_repo_issues_by_status(GitGud.Repo.t() | pos_integer(), keyword()) :: %{
  required(atom()) => non_neg_integer()
}

Returns the number of issues grouped by status for the given repo.

permissions(issue, user, repo_perms \\ nil)

Specs

permissions(GitGud.Issue.t(), GitGud.User.t() | nil, [atom()] | nil) :: [atom()]

Returns a list of permissions for the given issue and user.

repo_issue(repo, number, opts \\ [])

Specs

repo_issue(GitGud.Repo.t() | pos_integer(), pos_integer(), keyword()) ::
  GitGud.Issue.t() | nil

Returns a repository issue for the given number.

repo_issues(repo, opts \\ [])

Specs

repo_issues(GitGud.Repo.t() | pos_integer(), keyword()) :: [GitGud.Issue.t()]

Returns all issues for the given repo.

repo_issues_with_reply_count(repo, opts \\ [])

Specs

repo_issues_with_reply_count(GitGud.Repo.t() | pos_integer(), keyword()) :: [
  {GitGud.Issue.t(), pos_integer()}
]

Returns all issues with number of replies for the given repo.