GitGud.Mailer (git.limo v0.3.8)

Conveniences for composing and sending emails.

Functions

deliver(email)

Specs

deliver(any()) :: no_return()

deliver_later(email, opts \\ [])

deliver_later!(email, opts \\ [])

deliver_now(email, opts \\ [])

Specs

deliver_now(Bamboo.Email.t(), Enum.t()) ::
  {:ok, Bamboo.Email.t()}
  | {:ok, Bamboo.Email.t(), any()}
  | {:error, Exception.t() | String.t()}

deliver_now!(email, opts \\ [])

Specs

deliver_now!(Bamboo.Email.t(), Enum.t()) ::
  Bamboo.Email.t() | {Bamboo.Email.t(), any()}

password_reset_email(email)

Specs

password_reset_email(GitGud.Email.t()) :: Bamboo.Email.t()

Returns a password reset mail for the given email.

render(email, template, assigns \\ [])

Render an Phoenix template and set the body on the email.

Pass an atom as the template name (:welcome_email) to render HTML and plain text emails. Use a string if you only want to render one type, e.g. "welcome_email.text" or "welcome_email.html". Scroll to the top for more examples.

verification_email(email)

Specs

verification_email(GitGud.Email.t()) :: Bamboo.Email.t()

Returns a verification mail for the given email.