GitGud.Maintainer (git.limo v0.3.8)

Repository maintainer schema and helper functions.

A GitGud.Maintainer is used to grant GitGud.Repo permissions to a given GitGud.User.

Each repository maintainer also has a permission defining which actions he is able to perform on the repository. Following permissions are available:

:read -- can read and clone the repository.

:write -- can read, clone and push to the repository.

:admin -- can read, clone, push and administrate the repository.

By default, a newly created repository maintainer has :read permission.

Types

t()

Specs

t() :: %GitGud.Maintainer{
  __meta__: term(),
  id: term(),
  inserted_at: NaiveDateTime.t(),
  permission: binary(),
  repo: GitGud.Repo.t(),
  repo_id: pos_integer(),
  updated_at: NaiveDateTime.t(),
  user: GitGud.User.t(),
  user_id: pos_integer()
}

Functions

changeset(maintainer, params \\ %{})

Specs

changeset(t(), map()) :: Ecto.Changeset.t()

Returns a changeset for the given params.

delete(maintainer)

Specs

delete(t()) :: {:ok, t()} | {:error, Ecto.Changeset.t()}

delete(t()) :: t()

Deletes the given maintainer.

{:ok, maintainer} = GitGud.Maintainer.delete(maintainer)

delete!(maintainer)

Similar to delete!/1, but raises an Ecto.InvalidChangesetError if an error occurs.

update_permission(maintainer, permission)

Specs

update_permission(t(), binary()) :: {:ok, t()} | {:error, Ecto.Changeset.t()}

Updates the permission of the given maintainer.

{:ok, maintainer} = GitGud.Maintainer.update_permission(maintainer, :write)

This function validates the given permission using changeset/2.

update_permission(repo, user, permission)

Specs

update_permission(GitGud.Repo.t(), GitGud.User.t(), binary()) ::
  {:ok, t()} | :error

Updates the permission of the given user for the given repo.

{:ok, maintainer} = GitGud.Maintainer.update_permission(repo, user, :write)

This function validates the given permission using changeset/2.

update_permission!(maintainer, permission)

Specs

update_permission!(t(), binary()) :: t()

Similar to update_permission/2, but raises an Ecto.InvalidChangesetError if an error occurs.

update_permission!(repo, user, permission)

Specs

update_permission!(GitGud.Repo.t(), GitGud.User.t(), binary()) :: t()

Similar to update_permission/3, but raises an Ecto.NoResultsError if an error occurs.