Git Gud v0.1.0 GitGud.Repo

Git repository schema and helper functions.

A GitGud.Repo is used to create, update and delete Git repositories. It also provides a set of helper functions to run Git related functions.

Types

git_object()

git_object() ::
  GitGud.GitBlob.t()
  | GitGud.GitCommit.t()
  | GitGud.GitTag.t()
  | GitGud.GitTree.t()

git_revision()

git_revision() ::
  GitGud.GitReference.t() | GitGud.GitTag.t() | GitGud.GitCommit.t()

t()

t() :: %GitGud.Repo{
  __git__: term(),
  __meta__: term(),
  description: binary(),
  id: pos_integer(),
  inserted_at: NaiveDateTime.t(),
  maintainers: [GitGud.User.t()],
  name: binary(),
  owner: GitGud.User.t(),
  owner_id: pos_integer(),
  public: boolean(),
  pushed_at: NaiveDateTime.t(),
  updated_at: NaiveDateTime.t()
}

Functions

changeset(repo, params \\ %{})

changeset(t(), map()) :: Ecto.Changeset.t()

Returns a repository changeset for the given params.

create!(params, opts \\ [])

create!(map() | keyword(), keyword()) :: t()

Similar to create/2, but raises an Ecto.InvalidChangesetError if an error occurs.

create(params, opts \\ [])

create(map() | keyword(), keyword()) ::
  {:ok, t()} | {:error, Ecto.Changeset.t() | term()}

Creates a new repository.

{:ok, repo, git_handle} = GitGud.Repo.create(
  owner_id: user.id,
  name: "gitgud",
  description: "GitHub clone entirely written in Elixir.",
  public: true
)

This function validates the given params using changeset/2.

delete!(repo)

delete!(t()) :: t()

Similar to delete!/1, but raises an Ecto.InvalidChangesetError if an error occurs.

delete(repo)

delete(t()) :: {:ok, t()} | {:error, Ecto.Changeset.t()}

Deletes the given repo.

Repository associations (maintainers, issues, etc.) and related Git data will automatically be deleted.

empty?(repo)

empty?(t()) :: boolean()

Returns true if repo is empty; otherwhise returns false.

git_branch(repo, branch_name)

git_branch(t(), binary()) ::
  {:ok, GitGud.GitReference.t()} | {:error, term()}

Returns a Git branch for the given branch_name.

git_branches(repo)

git_branches(t()) :: {:ok, Stream.t()} | {:error, term()}

Returns all Git branches.

git_diff(old_tree, new_tree, opts \\ [])

Returns a diff with the difference between old_tree and new_tree.

git_head(repo)

git_head(t()) :: {:ok, GitGud.GitReference.t()} | {:error, term()}

Returns the Git reference pointed at by HEAD.

git_history(revision, opts \\ [])

Returns the commit history starting from the given revision.

git_object(repo, oid)

git_object(t(), binary()) :: {:ok, git_object()} | {:error, term()}

Returns the Git object matching the given oid.

git_push(repo, receive_pack)

git_push(t(), GitRekt.WireProtocol.ReceivePack.t()) :: :ok | {:error, term()}

Applies the given receive_pack command to the repo.

git_reference(repo, name_or_shorthand \\ :head)

git_reference(t(), binary()) ::
  {:ok, GitGud.GitReference.t()} | {:error, term()}

Returns the Git reference matching the given name_or_shorthand.

git_references(repo, glob \\ :undefined)

git_references(t(), binary() | :undefined) ::
  {:ok, Stream.t()} | {:error, term()}

Returns all Git references matching the given glob.

git_revision(repo, revision)

git_revision(t(), binary()) ::
  {:ok, git_object(), GitGud.GitReference.t() | nil} | {:error, term()}

Returns the Git object matching the given revision.

git_tag(repo, tag_name)

git_tag(t(), binary()) ::
  {:ok, GitGud.GitReference.t() | GitGud.GitTag.t()} | {:error, term()}

Returns a Git tag for the given tag_name.

git_tags(repo)

git_tags(t()) :: {:ok, Stream.t()} | {:error, term()}

Returns all Git tags.

git_tree(revision)

git_tree(git_revision()) :: {:ok, GitGud.GitTree.t()} | {:error, term()}

Returns the tree of the given revision.

maintainer(repo, user)

maintainer(t(), GitGud.User.t()) :: GitGud.Maintainer.t() | nil

Returns a single GitGud.Maintainer for the given repo and user.

maintainers(repo)

maintainers(t()) :: [GitGud.Maintainer.t()]

Returns the list of associated GitGud.Maintainer for the given repo.

open(repo)

open(t()) :: t()

Opens the Git repository for the given repo.

root_path()

root_path() :: Path.t() | nil

Returns the absolute path to the Git root directory.

update!(repo, params)

update!(t(), map() | keyword()) :: t()

Similar to update/2, but raises an Ecto.InvalidChangesetError if an error occurs.

update(repo, params)

update(t(), map() | keyword()) ::
  {:ok, t()} | {:error, Ecto.Changeset.t() | :file.posix()}

Updates the given repo with the given params.

{:ok, repo} = GitGud.Repo.update(repo, description: "Host open-source project without hassle.")

This function validates the given params using changeset/2.

workdir(repo)

workdir(t()) :: Path.t()

Returns the absolute path to the Git workdir for the given repo.

The path is a concatenation of root_path/0, repo.owner.login and repo.name.