GitGud.RepoQuery (git.limo v0.3.8)

Conveniences for repository related queries.

Types

user_param()

Specs

user_param() :: GitGud.User.t() | binary() | pos_integer()

Functions

by_id(id, opts \\ [])

Specs

by_id(pos_integer(), keyword()) :: GitGud.Repo.t() | nil

by_id([pos_integer()], keyword()) :: [GitGud.Repo.t()]

Returns a repository for the given id.

by_path(path, opts \\ [])

Specs

by_path(Path.t(), keyword()) :: GitGud.Repo.t() | nil

Returns a repository for the given path.

count_contributors(repo)

Specs

count_contributors(GitGud.Repo.t() | pos_integer()) :: non_neg_integer()

count_contributors([GitGud.Repo.t() | pos_integer()]) :: %{
  required(pos_integer()) => non_neg_integer()
}

Returns the number of contributors for the given repo.

issue_labels(repo, opts \\ [])

Specs

issue_labels(GitGud.Repo.t() | pos_integer(), keyword()) :: [
  GitGud.IssueLabel.t()
]

Returns a list of associated GitGud.IssueLabel for the given repo.

maintainer(repo_id, user)

Specs

maintainer(GitGud.Repo.t() | pos_integer(), GitGud.User.t()) ::
  GitGud.Maintainer.t() | nil

Returns a single GitGud.Maintainer for the given repo and user.

maintainers(repo)

Specs

maintainers(GitGud.Repo.t() | pos_integer()) :: [GitGud.Maintainer.t()]

Returns a list of associated GitGud.Maintainer for the given repo.

permissions(repo, user)

Specs

permissions(GitGud.Repo.t() | pos_integer(), GitGud.User.t() | nil) :: [atom()]

Returns a list of permissions for the given repo and user.

search(input, opts \\ [])

Specs

search(binary(), keyword()) :: [GitGud.User.t()]

Returns a list of users matching the given input.

user_repo(user, name, opts \\ [])

Specs

user_repo(user_param(), binary(), keyword()) :: GitGud.Repo.t() | nil

Returns a single repository for the given user and name.

user_repos(user, opts \\ [])

Specs

user_repos(user_param() | [user_param()], keyword()) :: [GitGud.Repo.t()]

Returns a list of repositories for the given user.