GitGud.RepoStorage (git.limo v0.3.8)

Conveniences for storing Git objects and meta objects.

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

cleanup(repo)

Specs

cleanup(GitGud.Repo.t()) :: {:ok, [Path.t()]} | {:error, term()}

Removes Git objects and references associated to the given repo.

ensure_volume_tagged()

Specs

ensure_volume_tagged() :: {:ok, binary()} | {:error, term()}

Ensures the associated volume is tagged.

init(repo, bare?)

Specs

init(GitGud.Repo.t(), boolean()) :: {:ok, GitRekt.Git.repo()} | {:error, term()}

Initializes a new Git repository for the given repo.

rename(old_repo, repo)

Specs

rename(GitGud.Repo.t(), GitGud.Repo.t()) :: {:ok, Path.t()} | {:error, term()}

Updates the workdir for the given repo.

start_link(volume, opts \\ [])

Specs

start_link(binary(), keyword()) :: GenServer.on_start()

Starts the storage server as part of a supervision tree.

volume()

Specs

volume() :: binary()

Returns the VOLUME identifier.

workdir(repo)

Specs

workdir(GitGud.Repo.t()) :: Path.t()

Returns the absolute path to the Git workdir for the given repo.

The path is a concatenation of the Git root path, repo.owner_login and repo.name.