GitGud.RepoSupervisor (git.limo v0.3.8)

Supervisor for dealing with repository processes.

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

register_volume(mod, volume)

Specs

register_volume(module(), binary() | nil) ::
  :ok | {:error, {:already_started, pid()}}

Registers the current process and volume with an unique name.

start_link(opts \\ [])

Specs

start_link(Supervisor.option() | Supervisor.init_option()) ::
  Supervisor.on_start()

Starts the supervisor as part of a supervision tree.

volume_name(mod, volume)

Specs

volume_name(module(), binary() | nil) :: term()

Returns a process name for a given module and volume.