GitGud.ReviewQuery (git.limo v0.3.8)

Conveniences for review related queries.

Functions

commit_line_review(id, opts \\ [])

Specs

commit_line_review(pos_integer(), keyword()) ::
  GitGud.CommitLineReview.t() | nil

Returns a commit line review for the given id.

commit_line_review(repo, commit, blob_oid, hunk, line, opts \\ [])

Specs

commit_line_review(
  GitGud.Repo.t() | pos_integer(),
  GitRekt.GitCommit.t() | GitRekt.Git.oid(),
  GitRekt.Git.oid(),
  non_neg_integer(),
  non_neg_integer(),
  keyword()
) :: GitGud.CommitLineReview.t() | nil

Returns a commit line review for the given repo, commit, blob_oid, hunk and line.

commit_line_reviews(repo, commit, opts \\ [])

Specs

commit_line_reviews(
  GitGud.Repo.t() | pos_integer(),
  GitRekt.GitCommit.t() | GitRekt.Git.oid(),
  keyword()
) :: [GitGud.CommitLineReview.t()]

Returns commit line reviews for the given repo and commit.

commit_line_reviews_comments(repo, commit, opts \\ [])

Specs

commit_line_reviews_comments(
  GitGud.Repo.t() | pos_integer(),
  GitRekt.GitCommit.t() | GitRekt.Git.oid(),
  keyword()
) :: [GitGud.CommitLineReview.t()]

Returns comments for the given repo and commit.

count_comments(repo, commit)

Specs

count_comments(
  Repot.t() | pos_integer(),
  GitRekt.GitCommit.t() | GitRekt.Git.oid()
) :: non_neg_integer()

count_comments(
  Repot.t() | pos_integer(),
  [GitRekt.GitCommit.t()] | [GitRekt.Git.oid()]
) :: {GitRekt.Git.oid(), non_neg_integer()}

Returns the number of comments for the given repo and commit.

count_comments_by_blob(repo, commit_oid)

Specs

count_comments_by_blob(
  GitGud.Repo.t() | pos_integer(),
  GitRekt.GitCommit.t() | GitRekt.Git.oid()
) :: %{required(binary()) => pos_integer()}

Returns the number of of comments grouped by blob oid for the given repo and commit.