GitGud.SSHKey (git.limo v0.3.8)

Secure Shell (SSH) authentication key schema and helper functions.

A GitGud.SSHKey is used essentially to authenticate SSH users via Public Key Authentication.

Types

t()

Specs

t() :: %GitGud.SSHKey{
  __meta__: term(),
  data: binary(),
  fingerprint: binary(),
  id: pos_integer(),
  inserted_at: NaiveDateTime.t(),
  name: binary(),
  used_at: NaiveDateTime.t(),
  user: GitGud.User.t(),
  user_id: pos_integer()
}

Functions

changeset(ssh_key, params \\ %{})

Specs

changeset(t(), map()) :: Ecto.Changeset.t()

Returns a changeset for the given key params.

If :name is omitted and the given public key contains a comment, the comment will be used as :name.

This function computes (see :public_key.ssh_decode/2 and :public_key.ssh_hostkey_fingerprint/1) and stores a fingerprint of the public key. It does not save the given :data into the database.

create(user, params)

Specs

create(GitGud.User.t(), map() | keyword()) ::
  {:ok, t()} | {:error, Ecto.Changeset.t()}

Creates a new SSH key with the given params.

{:ok, ssh_key} = GitGud.SSHKey.create(
  user,
  name: "My SSH Key",
  data: "ssh-rsa AAAAB3NzaC1yc2..."
)

This function validates the given params using changeset/2.

create!(user, params)

Specs

create!(GitGud.User.t(), map() | keyword()) :: t()

Similar to create/2, but raises an Ecto.InvalidChangesetError if an error occurs.

delete(ssh_key)

Specs

delete(t()) :: {:ok, t()} | {:error, Ecto.Changeset.t()}

Deletes the given ssh_key.

{:ok, ssh_key} = GitGud.SSHKey.delete(ssh_key)

delete!(ssh_key)

Specs

delete!(t()) :: t()

Similar to delete!/1, but raises an Ecto.InvalidChangesetError if an error occurs.

update_timestamp(ssh_key)

Specs

update_timestamp(t()) :: {:ok, t()} | {:error, Ecto.Changeset.t()}

Updates the timestamp for the the given ssh_key.

{:ok, ssh_key} = GitGud.SSHKey.update_timestamp(ssh_key)

update_timestamp!(ssh_key)

Specs

update_timestamp!(t()) :: t()

Similar to update_timestamp/1, but raises an Ecto.InvalidChangesetError if an error occurs.