GitGud.SSHServer (git.limo v0.3.8)

Secure Shell (SSH) server providing support for Git server commands.

The server handles following Git commands:

git-receive-pack - corresponding server-side command to git push.

git-upload-pack - corresponding server-side command to git fetch.

Example

To process Git commands over SSH, simply start the server as part of your supervision tree:

children = [
  Git.SSHServer
]

Supervisor.start_link(children, strategy: :one_for_one)

Authentication

A registered GitGud.User must authenticate with one of following methods:

public-key - if any of the associated GitGud.SSHKey matches.

password - if the given credentials are correct.

interactive - interactive login prompt allowing several tries.

See GitGud.Authorization for more details.

Types

t()

Specs

t() :: %GitGud.SSHServer{
  chan: :ssh.channel_id(),
  conn: :ssh.connection_ref(),
  repo: term(),
  request_size: non_neg_integer(),
  service: module(),
  user: GitGud.User.t()
}

Functions

child_spec(args)

Specs

child_spec([]) :: Supervisor.Spec.spec()

Returns a child-spec to use as part of a supervision tree.