GitGud.SmartHTTPBackend (git.limo v0.3.8)

Plug providing support for Git server commands over HTTP.

This plug handles following Git commands:

git-receive-pack - corresponding server-side command to git push.

git-upload-pack - corresponding server-side command to git fetch.

Example

Here is an example of GitGud.SmartHTTPBackend used in a Plug.Router to handle Git server commands:

defmodule SmartHTTPBackendRouter do
  use Plug.Router

  plug :match
  plug :fetch_query_params
  plug :dispatch

  get "/:user_login/:repo_name/info/refs", to: GitGud.SmartHTTPBackend, init_opts: :discover
  post "/:user_login/:repo_name/git-receive-pack", to: GitGud.SmartHTTPBackend, init_opts: :receive_pack
  post "/:user_login/:repo_name/git-upload-pack", to: GitGud.SmartHTTPBackend, init_opts: :upload_pack
end

Note that user_login and repo_name path parameters are mandatory.

To process Git commands over HTTP, simply start a Cowboy server as part of your supervision tree:

children = [
  {Plug.Cowboy, scheme: :http, plug: SmartHTTPBackendRouter}
]

Supervisor.start_link(children, strategy: :one_for_one)

Authentication

A registered GitGud.User can authenticate over HTTP via Basic Authentication. This is required to execute commands with granted permissions (such as pushing commits and cloning private repos).

See GitGud.Authorization for more details.

Functions

discover(conn, opts)

Specs

discover(Plug.Conn.t(), keyword()) :: Plug.Conn.t()

Returns all references available for the given Git repository.

receive_pack(conn, opts)

Specs

receive_pack(Plug.Conn.t(), keyword()) :: Plug.Conn.t()

Processes git-receive-pack requests.

upload_pack(conn, opts)

Specs

upload_pack(Plug.Conn.t(), keyword()) :: Plug.Conn.t()

Processes git-upload-pack requests.