GitGud.User (git.limo v0.3.8)

User account schema and helper functions.

A user is an individual's account that owns repositories and can make new content.

Types

t()

Specs

t() :: %GitGud.User{
  __meta__: term(),
  account: GitGud.Account.t(),
  avatar_url: binary(),
  bio: binary(),
  emails: [GitGud.Email.t()],
  gpg_keys: [GitGud.GPGKey.t()],
  id: pos_integer(),
  inserted_at: NaiveDateTime.t(),
  location: binary(),
  login: binary(),
  name: binary(),
  primary_email: GitGud.Email.t() | nil,
  primary_email_id: term(),
  public_email: GitGud.Email.t() | nil,
  public_email_id: term(),
  repos: [GitGud.Repo.t()],
  ssh_keys: [GitGud.SSHKey.t()],
  updated_at: NaiveDateTime.t(),
  website_url: binary()
}

Functions

create(params)

Specs

create(map() | keyword()) :: {:ok, t()} | {:error, Ecto.Changeset.t()}

Creates a new user with the given params.

{:ok, user} = GitGud.User.create(
  login: "redrabbit",
  name: "Mario Flach",
  emails: [
    %{address: "m.flach@almightycouch.com"}
  ],
  account: %{
    password: "qwertz"
  }
)

This function validates the given params using registration_changeset/2.

create!(params)

Specs

create!(map() | keyword()) :: t()

Similar to create/1, but raises an Ecto.InvalidChangesetError if an error occurs.

delete(user)

Specs

delete(t()) :: {:ok, t()} | {:error, Ecto.Changeset.t()}

Deletes the given user.

{:ok, user} = GitGud.User.delete(user)

User associations (repositories, comments, etc.) will automatically be deleted.

delete!(user)

Specs

delete!(t()) :: t()

Similar to delete!/1, but raises an Ecto.InvalidChangesetError if an error occurs.

email_changeset(user, email_type, email)

Specs

email_changeset(t(), :primary_email | :public_email, GitGud.Email.t()) ::
  Ecto.Changeset.t()

Returns a changeset for the given email.

oauth2_changeset(user, params \\ %{})

Specs

oauth2_changeset(t(), map()) :: Ecto.Changeset.t()

Returns a changeset for the given OAuth2.0 params.

See GitGud.Account.oauth2_changeset/2 for more details.

password_changeset(user, params \\ %{})

Specs

password_changeset(t(), map()) :: Ecto.Changeset.t()

Returns a changeset for the given password params.

See GitGud.Account.password_changeset/2 for more details.

profile_changeset(user, params \\ %{})

Specs

profile_changeset(t(), map()) :: Ecto.Changeset.t()

Returns a changeset for the given profile params.

registration_changeset(user, params \\ %{})

Specs

registration_changeset(t(), map()) :: Ecto.Changeset.t()

Returns a changeset for the given registration params.

update(user, changeset_type, params)

Specs

update(t(), atom(), struct() | map() | keyword()) ::
  {:ok, t()} | {:error, Ecto.Changeset.t()}

Updates the given user with the given changeset_type and params.

{:ok, user} = GitGud.User.update(user, :profile, name: "Mario Bros")

Following changeset types are available:

:profile -- see profile_changeset/2.

:password -- see password_changeset/2.

:oauth2 -- see oauth2_changeset/2.

This function can also be used to update email associations, for example:

{:ok, user} = GitGud.User.update(user, :primary_email, email)

update!(user, changeset_type, params)

Specs

update!(t(), atom(), struct() | map() | keyword()) :: t()

Similar to update/3, but raises an Ecto.InvalidChangesetError if an error occurs.

verified?(user)

Specs

verified?(t()) :: boolean()

Returns true is user is verified; otherwise returns false.