Git Gud v0.1.0 GitGud.Web.CodebaseController

Module responsible for CRUD actions on GitGud.Repo.

Functions

blob(conn, params)

blob(Plug.Conn.t(), map()) :: Plug.Conn.t()

Renders a blob for a specific revision and path.

branches(conn, params)

branches(Plug.Conn.t(), map()) :: Plug.Conn.t()

Renders all branches of a repository.

commit(conn, params)

commit(Plug.Conn.t(), map()) :: Plug.Conn.t()

Renders a single commit.

history(conn, params)

history(Plug.Conn.t(), map()) :: Plug.Conn.t()

Renders all commits for a specific revision.

show(conn, params)

show(Plug.Conn.t(), map()) :: Plug.Conn.t()

Renders a repository codebase overview.

tags(conn, params)

tags(Plug.Conn.t(), map()) :: Plug.Conn.t()

Renders all tags of a repository.

tree(conn, params)

tree(Plug.Conn.t(), map()) :: Plug.Conn.t()

Renders a tree for a specific revision and path.