Git Gud v0.1.0 GitGud.Web.ErrorView

Module providing error views for most common errors.

Functions

__phoenix_recompile__?()

Returns true whenever the list of templates changes in the filesystem.

__resource__()

The resource name, as an atom, for this view

__templates__()

Returns the template root alongside all templates.

render(template, assigns \\ %{})

Renders the given template locally.

template_not_found(template, assigns)

template_not_found(Phoenix.Template.name(), map()) :: no_return()

Callback invoked when no template is found. By default it raises but can be customized to render a particular template.