GitGud.Web.Markdown (git.limo v0.3.8)

Conveniences for rendering Markdown.

Functions

markdown(content, opts \\ [])

Specs

markdown(binary() | nil, keyword() | map()) :: binary() | nil

Renders a Markdown formatted content to HTML.

markdown_safe(content, opts \\ [])

Specs

markdown_safe(binary() | nil, keyword() | map()) :: binary() | nil

Renders a Markdown formatted content to HTML and marks it as safe for Phoenix to render.

parse(content)

Specs

parse(binary() | nil) :: map()

Returns a map of potential matches for the given content.