GitGud.Web.ReactComponents (git.limo v0.3.8)

Functions to make rendering React components.

Functions

react_component(name)

Specs

react_component(binary()) :: iodata()

Generates a :div containing the named React component with no props or attrs.

Returns safe html: {:safe, [60, "div", ...]}.

You can utilize this in your Phoenix views:

<%= GitGud.Web.React.react_component("MyComponent") %>

The resulting <div> tag is formatted specifically for the included javascript helper to then turn into your named React component.

react_component(name, props)

Specs

react_component(binary(), Keyword.t()) :: iodata()

Generates a :div containing the named React component with the given props.

Returns safe html: {:safe, [60, "div", ...]}.

You can utilize this in your Phoenix views:

<%= GitGud.Web.React.react_component("MyComponent", %{language: "elixir", awesome: true}) %>

The resulting <div> tag is formatted specifically for the included javascript helper to then turn into your named React component and then pass in the props specified.

react_component(name, props, attrs)

Specs

react_component(binary(), Keyword.t(), Keyword.t()) :: iodata()

Generates a :div containing the named React component with the given props and attrs.

Returns safe html: {:safe, [60, "div", ...]}.

You can utilize this in your Phoenix views:

<%= GitGud.Web.React.react_component("MyComponent", %{language: "elixir", awesome: true}, class: "my-component") %>

The resulting <div> tag is formatted specifically for the included javascript helper to then turn into your named React component and then pass in the props specified.

react_component(name, props, attrs, inner_block)

Generates a :div containing the named React component with the given props and attrs.

Returns safe html: {:safe, [60, "div", ...]}.

You can utilize this in your Phoenix views:

<%= GitGud.Web.React.react_component("MyComponent", %{language: "elixir", awesome: true}, class: "my-component") do %>
  <p>
    <em>this text is rendered until the component is initialized</em>
  </p>
<% end %>

The resulting <div> tag is formatted specifically for the included javascript helper to then turn into your named React component and then pass in the props specified.