GitRekt.Cache behaviour (git.limo v0.3.8)

Behaviour for caching Git operations.

Types

cache()

Specs

cache() :: term()

cache_entry()

Specs

cache_entry() :: term()

cache_key()

Specs

cache_key() :: term()

op()

Specs

op() :: atom() | tuple()

Callbacks

fetch_cache(cache, cache_key)

Specs

fetch_cache(cache(), cache_key()) :: cache_entry() | nil

Fetches the value for a specific cache_key.

init_cache(path, opts)

Specs

init_cache(path :: Path.t(), opts :: keyword()) :: cache()

Initialize the cache for the given path.

make_cache_key(op)

Specs

make_cache_key(op()) :: cache_key() | nil

Returns the cache key for the given op or nil if the operation should not be cached.

put_cache(cache, cache_key, cache_entry)

Specs

put_cache(cache(), cache_key(), cache_entry()) :: :ok

Puts the given cache_entry under cache_key.