GitRekt.Git (git.limo v0.3.8)

Erlang NIF that exposes a subset of libgit2's library functions.

Most functions available in this module are implemented in C for performance reasons. These functions are compiled into a dynamic loadable, shared library. They are called like any other Elixir functions.

As a NIF library is dynamically linked into the emulator process, this is the fastest way of calling C-code from Erlang (alongside port drivers). Calling NIFs requires no context switches. But it is also the least safe, because a crash in a NIF brings the emulator down too.

Erlang documentation - NIFs

Example

Let's start with a basic code example showing the last commit author and message:

alias GitRekt.Git

# load repository
{:ok, repo} = Git.repository_open("/tmp/my-repo.git")

# fetch commit pointed by master
{:ok, :commit, _oid, commit} = Git.reference_peel(repo, "refs/heads/master")

# fetch commit author & message
{:ok, name, email, time, _offset} = Git.commit_author(commit)
{:ok, message} = Git.commit_message(commit)

IO.puts "Last commit by #{name} <#{email}>:"
IO.puts message

First we open our repository using repository_open/1, passing the path of the Git repository. We can fetch a commit by passing the exact reference path to reference_peel/2. In our example, this allows us to retrieve the commit refs/heads/master is pointing to.

This is one of many ways to fetch a given revision, reference_lookup/2 and reference_glob/2 offer similar functionalities. There are other related functions such as revparse_single/2 and revparse_ext/2 which provide support for parsing revspecs.

Walk commit history

In order to walk the commit ancestry chain, we have a few functions at our disposal: revwalk_new/1, revwalk_push/2, revwalk_next/1, revwalk_reset/1, etc.

# create revision walk iterator
{:ok, revwalk} = Git.revwalk_new(repo)

# set root commit for traversal
 :ok = Git.revwalk_push(walk, commit_oid)

# create (lazy) stream of ancestors from iterator
{:ok, stream} = Git.revwalk_stream(walk)

for ancestor_oid <- stream do
  # fetch commit object
  {:ok, :commit, commit} = Git.object_lookup(ancestor_oid)

  # fetch commit message
  {:ok, message} = Git.commit_message(commit)

  IO.puts "#{Git.oid_fmt_short(ancestor_oid)} - #{message}"
end

In this example revwalk_new/1 returns a revwalk/0, a mutable C-like iteratable object. This means that revwalk_push/2 mutates the revwalk object instead of returning a new object.

The revwalk_stream/1 function converts the revwalk iterator to a Enumerable.t/0 we can then use to walk the commit ancestry chain.

When iterating through a commit's history, revwalk_sorting/2 and revwalk_simplify_first_parent/1 provide conveniences for sorting and filtering while revwalk_push/3 can be used to hide specific commits.

Retrieve blobs & trees

In order to access the actual files and directories of a repository, we have to retrieve the Git blob and tree objects of a given revision. Here we simply list files and folders at the root directory:

# fetch commit tree
{:ok, tree} = Git.commit_tree(commit)

# fetch tree entries at /
{:ok, tree_entries} = Git.tree_entries(tree)

for {mode, type, oid, name} <- tree_entries do
  # fetch tree entry object by oid (blob or tree)
  case Git.object_lookup(repo, oid) do
    {:ok, :blob, blob} ->
      # fetch blob size
      {:ok, blob_size} = Git.blob_size(blob)
      IO.puts "#{name} -- #{blob_size} bytes"
    {:ok, :tree, tree} ->
      # fetch number of sub entries
      {:ok, count} <- Git.tree_count(tree)
      IO.puts "#{name}/ -- #{count} items"
  end
end

Note that tree_entries/1 and tree_nth/2 return a tuple in the form of {mode, type, oid, name}. In order to call blob and tree specific functions such as blob_size/1 and tree_count/1, we still need to lookup the Git object using object_lookup/2.

Here's an other example showing a convenient way to retrieve a tree entry by path:

# fetch commit tree
{:ok, tree} = Git.commit_tree(commit)

# fetch blob by path
{:ok, mode, :blob, oid, name} Git.tree_bypath(tree, "README.md")

# fetch blob object by oid
{:ok, :blob, blob} = Git.object_lookup(repo, oid)

# fetch blob content
{:ok, data} = Git.blob_content(blob)

IO.binwrite data

Compare revisions

Now that we know how to access files and directories, it might be interesting to determine the changes between two versions. In order to do so, we need to compare two tree objects from different revisions:

# fetch tree for tag v1.2
{:ok, v1_2, :commit, _oid} = Git.revparse_single(repo, "v1.2")
{:ok, v1_2_tree} = Git.commit_tree(v1_2)

# fetch tree for tag v1.3
{:ok, v1_3, :commit, _oid} = Git.revparse_single(repo, "v1.3")
{:ok, v1_3_tree} = Git.commit_tree(v1_3)

# fetch diff between v1.2 and v1.3
{:ok, diff} = Git.diff_tree(repo, v1_2_tree, v1_3_tree)

# format diff to string
{:ok, patch} = Git.diff_format(diff, :patch)

IO.puts patch

Note that diff_tree/4 takes options such as :pathspec allowing to filter changes based on a given path.

For example, we might want to see the different between two revision for a specific file. In this case we could modify the above example as follow to print changes affecting README.md:

# fetch diff between v1.2 and v1.3 for README.md
{:ok, diff} = Git.diff_tree(repo, v1_2_tree, v1_3_tree, pathspec: "README.md")

# format diff to string
{:ok, patch} = Git.diff_format(diff, :patch)

IO.puts patch

Commit changes

Committing changes to a repository is done in a serie of distinct steps.

First, we add a new blob object to the repository:

# Blob content
blob_content = "Hello world\n"

# Open repository ODB
{:ok, odb} = Git.repository_get_odb(repo)

# Write new blob object
{:ok, blob_oid} = Git.odb_write(odb, blob_content, :blob)

We create an in-memory index to stage our modifications (an index is a list of path names, each with permissions and the SHA1 of a blob object). In order to create a new tree object reflecting our changes, we have to assign our new blob object at a given path in the index and write the index back to the repository:

# Blob path
blob_path = "README"

# Create new index
{:ok, index} = Git.index_new()

# Read last commit tree into index
:ok = Git.index_read_tree(index, tree)

# Add newly added blob object to index
:ok = Git.index_add(repo, index, blob_oid, blob_path, byte_size(blob_content), 0o100644)

# Write index
{:ok, tree_oid} = Git.index_write_tree(index)

The repository now contains a new blob object and a new tree object reflecting our changes. We now have all the ingredients for creating a commit and update the master branch accordingly:

# Commit message
commit_message = "Add README"

# Fetch repository default signature for authoring and committing
{:ok, sig_name, sig_email, sig_ts, sig_tz} = Git.signature_default(repo)
commit_author = {sig_name, sig_email, sig_ts, sig_tz}
commit_committer = commit_author

# Fetch reference to update
{:ok, :commit, parent_oid, _parent} = Git.reference_peel(repo, "refs/heads/master")

# Create new commit
{:ok, commit_oid} = Git.commit_create(repo, :undefined, commit_author, commit_committer, :undefined, commit_message, tree_oid, [parent_oid])

# Update master branch to new commit
:ok = Git.reference_create(repo, "refs/heads/master", :oid, commit_oid)

IO.puts "File #{blob_path} added in commit #{Git.oid_fmt(commit_oid)}."

We have created a commit pointing at the new tree object. The commit refers the newly created tree and requires two user signatures (author and committer), a commit message and the commit ancestor(s).

Finally we have updated the master branch to point at our new commit.

Thread safety

Accessing a repo/0 or any NIF allocated pointer (blob/0, commit/0, config/0, etc.) from multiple processes simultaneously is not safe. These pointers should never be shared across processes.

In order to access a repository in a concurrent manner, each process has to initialize it's own repository resource using repository_open/1. Alternatively, the GitRekt.GitAgent module provides a similar API but can use a dedicated process, so that its access can be serialized.

Types

blob()

Specs

blob() :: reference()

commit()

Specs

commit() :: reference()

config()

Specs

config() :: reference()

diff()

Specs

diff() :: reference()

diff_delta()

Specs

diff_delta() :: {diff_file(), diff_file(), non_neg_integer(), non_neg_integer()}

diff_file()

Specs

diff_file() :: {oid(), binary(), integer(), non_neg_integer()}

diff_format()

Specs

diff_format() :: :patch | :patch_header | :raw | :name_only | :name_status

diff_hunk()

Specs

diff_hunk() :: {binary(), integer(), integer(), integer(), integer()}

diff_line()

Specs

diff_line() :: {char(), integer(), integer(), integer(), integer(), binary()}

index()

Specs

index() :: reference()

index_entry()

Specs

index_entry() ::
  {integer(), integer(), non_neg_integer(), non_neg_integer(),
   non_neg_integer(), non_neg_integer(), non_neg_integer(), integer(), binary(),
   non_neg_integer(), non_neg_integer(), binary()}

indexer_progress()

Specs

indexer_progress() :: reference()

obj()

Specs

obj() :: blob() | commit() | tree() | tag()

obj_type()

Specs

obj_type() :: :blob | :commit | :tree | :tag

odb()

Specs

odb() :: reference()

odb_type()

Specs

odb_type() :: atom()

odb_writepack()

Specs

odb_writepack() :: reference()

odb_writepack_progress()

Specs

odb_writepack_progress() :: map()

oid()

Specs

oid() :: binary()

pack()

Specs

pack() :: reference()

ref_iter()

Specs

ref_iter() :: reference()

ref_type()

Specs

ref_type() :: :oid | :symbolic

reflog_entry()

Specs

reflog_entry() ::
  {binary(), binary(), non_neg_integer(), non_neg_integer(), oid(), oid(),
   binary()}

repo()

Specs

repo() :: reference()

revwalk()

Specs

revwalk() :: reference()

revwalk_sort()

Specs

revwalk_sort() :: :sort_topo | :sort_time | :sort_reverse

signature()

Specs

signature() :: {binary(), binary(), non_neg_integer(), non_neg_integer()}

tag()

Specs

tag() :: reference()

tree()

Specs

tree() :: reference()

tree_entry()

Specs

tree_entry() :: {integer(), :blob | :tree, oid(), binary()}

worktree()

Specs

worktree() :: reference()

Functions

blob_content(blob)

Specs

blob_content(blob()) :: {:ok, binary()} | {:error, term()}

Returns the raw content of the given blob.

blob_size(blob)

Specs

blob_size(blob()) :: {:ok, integer()} | {:error, term()}

Returns the size in bytes of the given blob.

commit_author(commit)

Specs

commit_author(commit()) ::
  {:ok, binary(), binary(), non_neg_integer(), non_neg_integer()}
  | {:error, term()}

Returns the author of the given commit.

commit_committer(commit)

Specs

commit_committer(commit()) ::
  {:ok, binary(), binary(), non_neg_integer(), non_neg_integer()}
  | {:error, term()}

Returns the committer of the given commit.

commit_create(repo, ref, author, commiter, encoding, message, tree, parents)

Specs

commit_create(
  repo(),
  binary() | :undefined,
  signature(),
  signature(),
  binary() | :undefined,
  binary(),
  oid(),
  [binary()]
) :: {:ok, oid()} | {:error, term()}

Creates a new commit with the given params.

commit_header(commit, field)

Specs

commit_header(commit(), binary()) :: {:ok, binary()} | {:error, term()}

Returns an arbitrary header field of the given commit.

commit_message(commit)

Specs

commit_message(commit()) :: {:ok, binary()} | {:error, term()}

Returns the message for the given commit.

commit_parent(commit, index)

Specs

commit_parent(commit(), non_neg_integer()) ::
  {:ok, oid(), commit()} | {:error, term()}

Looks for a parent commit of the given commit by its index.

commit_parent_count(commit)

Specs

commit_parent_count(commit()) :: oid()

Returns the number of parents for the given commit.

commit_parents(commit)

Specs

commit_parents(commit()) :: {:ok, Enumerable.t()} | {:error, term()}

Returns parent commits of the given commit.

commit_raw_header(commit)

Specs

commit_raw_header(commit()) :: {:ok, binary()} | {:error, term()}

Returns the full raw header of the given commit.

commit_time(commit)

Specs

commit_time(commit()) ::
  {:ok, non_neg_integer(), non_neg_integer()} | {:error, term()}

Returns the time of the given commit.

commit_tree(commit)

Specs

commit_tree(commit()) :: {:ok, oid(), tree()} | {:error, term()}

Returns the tree for the given commit.

commit_tree_id(commit)

Specs

commit_tree_id(commit()) :: oid()

Returns the tree id for the given commit.

config_get_bool(config, name)

Specs

config_get_bool(config(), binary()) :: {:ok, boolean()} | {:error, term()}

Returns the value of the config entry with the given name.

config_get_string(config, name)

Specs

config_get_string(config(), binary()) :: {:ok, binary()} | {:error, term()}

Returns the value of the config entry with the given name.

config_open(path)

Specs

config_open(binary()) :: {:ok, config()} | {:error, term()}

Returns a config handle for the given path.

config_set_bool(config, name, val)

Specs

config_set_bool(config(), binary(), boolean()) :: :ok | {:error, term()}

Sets the config entry with the given name to val.

config_set_string(config, name, val)

Specs

config_set_string(config(), binary(), binary()) :: :ok | {:error, term()}

Sets the config entry with the given name to val.

diff_delta_count(diff)

Specs

diff_delta_count(diff()) :: {:ok, non_neg_integer()} | {:error, term()}

Returns the number of deltas in the given diff.

diff_deltas(diff)

Specs

diff_deltas(diff()) ::
  {:ok, [{diff_delta(), [{diff_hunk(), [diff_line()]}]}]} | {:error, term()}

Returns a list of deltas for the given diff.

diff_format(diff, format \\ :patch)

Specs

diff_format(diff(), diff_format()) :: {:ok, binary()} | {:error, term()}

Returns a binary represention of the given diff.

diff_stats(diff)

Specs

diff_stats(diff()) ::
  {:ok, non_neg_integer(), non_neg_integer(), non_neg_integer()}
  | {:error, term()}

Returns stats for the given diff.

diff_tree(repo, old_tree, new_tree, opts \\ [])

Returns a diff with the difference between two tree objects.

graph_ahead_behind(repo, local, upstream)

Specs

graph_ahead_behind(repo(), oid(), oid()) ::
  {:ok, non_neg_integer(), non_neg_integer()}

Returns the number of unique commits between two commit objects.

index_add(index, entry)

Specs

index_add(index(), index_entry()) :: :ok | {:error, term()}

Adds or updates the given entry.

index_bypath(index, path, stage)

Specs

index_bypath(index(), Path.t(), non_neg_integer()) ::
  {:ok, integer(), integer(), non_neg_integer(), non_neg_integer(),
   non_neg_integer(), non_neg_integer(), non_neg_integer(), integer(), binary(),
   non_neg_integer(), non_neg_integer(), binary()}
  | {:error, term()}

Retrieves an entry contained in the index given its relative path.

index_clear(index)

Specs

index_clear(index()) :: :ok | {:error, term()}

Clears the contents (all the entries) of the given index.

index_count(index)

Specs

index_count(index()) :: non_neg_integer()

Returns the number of entries in the given index.

index_new()

Specs

index_new() :: {:ok, index()} | {:error, term()}

Creates an new in-memory index object.

index_nth(index, nth)

Specs

index_nth(index(), non_neg_integer()) ::
  {:ok, integer(), integer(), non_neg_integer(), non_neg_integer(),
   non_neg_integer(), non_neg_integer(), non_neg_integer(), integer(), binary(),
   non_neg_integer(), non_neg_integer(), binary()}
  | {:error, term()}

Looks for an entry by its position in the given index.

index_read_tree(index, tree)

Specs

index_read_tree(index(), tree()) :: :ok | {:error, term()}

Reads the given tree into the given index file with stats.

index_remove(index, path, stage \\ 0)

Specs

index_remove(index(), Path.t(), non_neg_integer()) :: :ok | {:error, term()}

Removes an entry from the given index.

index_remove_dir(index, path, stage \\ 0)

Specs

index_remove_dir(index(), Path.t(), non_neg_integer()) :: :ok | {:error, term()}

Removes all entries from the given index under a given directory.

index_write(index)

Specs

index_write(index()) :: :ok | {:error, term()}

Writes the given index from memory back to disk using an atomic file lock.

index_write_tree(index)

Specs

index_write_tree(index()) :: {:ok, oid()} | {:error, term()}

Writes the given index as a tree.

index_write_tree(index, repo)

Specs

index_write_tree(index(), repo()) :: {:ok, oid()} | {:error, term()}

Writes the given index as a tree.

library_version()

Specs

library_version() :: {integer(), integer(), integer()}

Returns the libgit2 library version.

object_id(obj)

Specs

object_id(obj()) :: {:ok, oid()} | {:error, term()}

Returns the OID for the given obj.

object_lookup(repo, oid)

Specs

object_lookup(repo(), oid()) :: {:ok, obj_type(), obj()} | {:error, term()}

Looks for an object with the given oid.

object_repository(obj)

Specs

object_repository(obj()) :: {:ok, repo()} | {:error, term()}

Returns the repository that owns the given obj.

object_zlib_inflate(data, buffer_size \\ 16384)

Specs

object_zlib_inflate(binary(), pos_integer()) ::
  {:ok, iodata(), non_neg_integer()} | {:error, term()}

Inflates the given data with zlib.

odb_get_writepack(odb)

Specs

odb_get_writepack(odb()) :: {:ok, odb_writepack()} | {:error, term()}

Returns an ODB write-pack for the given odb.

odb_object_exists?(odb, oid)

Specs

odb_object_exists?(odb(), oid()) :: boolean()

Returns true if the given oid exists in odb; elsewise returns false.

odb_object_hash(type, data)

Specs

odb_object_hash(obj_type(), binary()) :: {:ok, oid()} | {:error, term()}

Returns the OID of an object type and raw data.

The resulting SHA-1 OID will be the identifier for the data buffer as if the data buffer it were to written to the ODB.

odb_read(odb, oid)

Specs

odb_read(odb(), oid()) :: {:ok, obj_type(), binary()}

Return the uncompressed, raw data of an ODB object.

odb_write(odb, data, type)

Specs

odb_write(odb(), binary(), odb_type()) :: {:ok, oid()} | {:error, term()}

Writes the given object data with the given type into the odb.

odb_write_pack(odb, data)

Specs

odb_write_pack(odb(), binary()) :: :ok | {:error, term()}

Writes the given PACK data into the odb.

odb_writepack_append(odb_writepack, data, progress)

Specs

odb_writepack_append(odb_writepack(), binary(), indexer_progress()) ::
  :ok | {:error, term()}

Appends the given data to the odb_writepack.

odb_writepack_commit(odb_writepack, progress)

Specs

odb_writepack_commit(odb_writepack(), indexer_progress()) ::
  :ok | {:error, term()}

Commits the written data to the odb_writepack.

oid_fmt(oid)

Specs

oid_fmt(oid()) :: binary()

Returns the SHA hash for the given oid.

oid_fmt_short(oid)

Specs

oid_fmt_short(oid()) :: binary()

Returns the abbreviated SHA hash for the given oid.

oid_parse(hash)

Specs

oid_parse(binary()) :: oid()

Returns the OID for the given SHA hash.

pack_data(pack)

Specs

pack_data(pack()) :: {:ok, binary()} | {:error, term()}

Returns a PACK file for the given pack.

pack_insert_commit(pack, oid)

Specs

pack_insert_commit(pack(), oid()) :: :ok | {:error, term()}

Inserts commit as well as the completed referenced tree.

pack_insert_walk(pack, walk)

Specs

pack_insert_walk(pack(), revwalk()) :: :ok | {:error, term()}

Inserts objects as given by the walk.

pack_new(repo)

Specs

pack_new(repo()) :: {:ok, pack()} | {:error, term()}

Creates a new PACK object for the given repo.

pathspec_match_tree(tree, pathspec)

Specs

pathspec_match_tree(tree(), [binary()]) :: {:ok, boolean()} | {:error, term()}

Returns true if tree matches the given pathspec; otherwise returns false.

reference_create(repo, name, type, target, force \\ false)

Specs

reference_create(repo(), binary(), ref_type(), binary() | oid(), boolean()) ::
  :ok | {:error, term()}

Creates a new reference name which points to an object or to an other reference.

reference_delete(repo, name)

Specs

reference_delete(repo(), binary()) :: :ok | {:error, term()}

Deletes an existing reference.

reference_dwim(repo, short_name)

Specs

reference_dwim(repo(), binary()) ::
  {:ok, binary(), ref_type(), binary()} | {:error, term()}

Looks for a reference by DWIMing its short_name.

reference_glob(repo, glob)

Specs

reference_glob(repo(), binary()) :: {:ok, [binary()]} | {:error, term()}

Similar to reference_list/1 but allows glob patterns.

reference_iterator(repo, glob \\ :undefined)

Specs

reference_iterator(repo(), binary() | :undefined) ::
  {:ok, ref_iter()} | {:error, term()}

Returns an iterator for the references that match the specific glob pattern.

reference_list(repo)

Specs

reference_list(repo()) :: {:ok, [binary()]} | {:error, term()}

Returns all references for the given repo.

reference_log?(repo, name)

Specs

reference_log?(repo(), binary()) :: {:ok, boolean()} | {:error, term()}

Returns true if a reflog exists for the given reference name.

reference_lookup(repo, name)

Specs

reference_lookup(repo(), binary()) ::
  {:ok, binary(), ref_type(), binary()} | {:error, term()}

Looks for a reference by name.

reference_next(iter)

Specs

reference_next(ref_iter()) ::
  {:ok, binary(), binary(), ref_type(), binary()} | {:error, term()}

Returns the next reference.

reference_peel(repo, name, type \\ :undefined)

Specs

reference_peel(repo(), binary(), obj_type() | :undefined) ::
  {:ok, obj_type(), oid(), obj()} | {:error, term()}

Recursively peels the given reference name until an object of type type is found.

reference_resolve(repo, name)

Specs

reference_resolve(repo(), binary()) ::
  {:ok, binary(), binary(), oid()} | {:error, term()}

Resolves a symbolic reference to a direct reference.

reference_stream(repo, glob \\ :undefined)

Specs

reference_stream(repo(), binary() | :undefined) ::
  {:ok, Enumerable.t()} | {:error, term()}

Returns a stream for the references that match the specific glob pattern.

reference_to_id(repo, name)

Specs

reference_to_id(repo(), binary()) :: {:ok, oid()} | {:error, term()}

Looks for a reference by name and returns its id.

reflog_count(repo, name)

Specs

reflog_count(repo(), binary()) :: {:ok, pos_integer()} | {:error, term()}

Reads the number of entry for the given reflog name.

reflog_delete(repo, name)

Specs

reflog_delete(repo(), binary()) :: :ok | {:error, term()}

Deletes the reflog for the given reference name.

reflog_read(repo, name)

Specs

reflog_read(repo(), binary()) :: {:ok, [reflog_entry()]} | {:error, term()}

Reads the reflog for the given reference name.

repository_bare?(repo)

Specs

repository_bare?(repo()) :: boolean()

Returns true if repo is bare; elsewise returns false.

repository_discover(path)

Specs

repository_discover(Path.t()) :: {:ok, Path.t()} | {:error, term()}

Looks for a repository and returns its path.

repository_empty?(repo)

Specs

repository_empty?(repo()) :: boolean()

Returns true if repo is empty; elsewise returns false.

repository_get_config(repo)

Specs

repository_get_config(repo()) :: {:ok, config()} | {:error, term()}

Returns the config for the given repo.

repository_get_index(repo)

Specs

repository_get_index(repo()) :: {:ok, index()} | {:error, term()}

Returns the index for the given repository.

repository_get_odb(repo)

Specs

repository_get_odb(repo()) :: {:ok, odb()} | {:error, term()}

Returns the ODB for the given repository.

repository_get_path(repo)

Specs

repository_get_path(repo()) :: Path.t()

Returns the absolute path for the given repo.

repository_get_workdir(repo)

Specs

repository_get_workdir(repo()) :: Path.t()

Returns the normalized path to the working directory for the given repo.

repository_init(path, bare \\ false)

Specs

repository_init(Path.t(), boolean()) :: {:ok, repo()} | {:error, term()}

Initializes a new repository at the given path.

repository_open(path)

Specs

repository_open(Path.t()) :: {:ok, repo()} | {:error, term()}

Returns a repository handle for the path.

revparse_ext(repo, revision)

Specs

revparse_ext(repo(), binary()) ::
  {:ok, obj(), obj_type(), oid(), binary() | nil} | {:error, term()}

Finds a single object and intermediate reference, as specified by the given revision.

revparse_single(repo, revision)

Specs

revparse_single(repo(), binary()) ::
  {:ok, obj(), obj_type(), oid()} | {:error, term()}

Finds a single object, as specified by the given revision.

revwalk_new(repo)

Specs

revwalk_new(repo()) :: {:ok, reference()} | {:error, term()}

Creates a new revision walk object for the given repo.

revwalk_next(walk)

Specs

revwalk_next(revwalk()) :: {:ok, oid()} | {:error, term()}

Returns the next commit from the given revision walk.

revwalk_pack(walk)

Specs

revwalk_pack(revwalk()) :: {:ok, binary()} | {:error, term()}

Returns a PACK file for the given walk.

revwalk_push(walk, oid, hide \\ false)

Specs

revwalk_push(revwalk(), oid(), boolean()) :: :ok | {:error, term()}

Adds a new root for the traversal.

revwalk_repository(walk)

Specs

revwalk_repository(revwalk()) :: {:ok, repo()} | {:error, term()}

Returns the repository on which the given walker is operating.

revwalk_reset(walk)

Specs

revwalk_reset(revwalk()) :: revwalk()

Resets the revision walk for reuse.

revwalk_simplify_first_parent(walk)

Specs

revwalk_simplify_first_parent(revwalk()) :: :ok | {:error, term()}

Simplifies the history by first-parent.

revwalk_sorting(walk, sort_mode)

Specs

revwalk_sorting(revwalk(), [revwalk_sort()]) :: :ok | {:error, term()}

Changes the sorting mode when iterating through the repository's contents.

revwalk_stream(walk)

Specs

revwalk_stream(revwalk()) :: {:ok, Enumerable.t()} | {:error, term()}

Returns a stream for the given revision walk.

signature_default(repo)

Specs

signature_default(repo()) ::
  {:ok, binary(), binary(), non_neg_integer(), non_neg_integer()}
  | {:error, term()}

Returns the default signature for the given repo.

signature_new(name, email)

Specs

signature_new(binary(), binary()) :: {:ok, binary(), binary()}

Creates a new signature with the given name and email.

signature_new(name, email, time)

Specs

signature_new(binary(), binary(), non_neg_integer()) ::
  {:ok, binary(), binary(), non_neg_integer(), non_neg_integer()}
  | {:error, term()}

Creates a new signature with the given name, email and time.

tag_author(tag)

Specs

tag_author(tag()) ::
  {:ok, binary(), binary(), non_neg_integer(), non_neg_integer()}
  | {:error, term()}

Returns the author of the given tag.

tag_list(repo)

Specs

tag_list(repo()) :: {:ok, [binary()]} | {:error, term()}

Returns all tags for the given repo.

tag_message(tag)

Specs

tag_message(tag()) :: {:ok, binary()} | {:error, term()}

Returns the message of the given tag.

tag_name(tag)

Specs

tag_name(tag()) :: {:ok, binary()} | {:error, term()}

Returns the name of the given tag.

tag_peel(tag)

Specs

tag_peel(tag()) :: {:ok, obj_type(), oid(), obj()} | {:error, term()}

Recursively peels the given tag until a non tag object is found.

tree_byid(tree, oid)

Specs

tree_byid(tree(), oid()) ::
  {:ok, integer(), atom(), binary(), binary()} | {:error, term()}

Retrieves a tree entry owned by the given tree, given its id.

tree_bypath(tree, path)

Specs

tree_bypath(tree(), Path.t()) ::
  {:ok, integer(), atom(), binary(), binary()} | {:error, term()}

Retrieves a tree entry contained in the given tree or in any of its subtrees, given its relative path.

tree_count(tree)

Specs

tree_count(tree()) :: {:ok, non_neg_integer()} | {:error, term()}

Returns the number of entries listed in the given tree.

tree_entries(tree)

Specs

tree_entries(tree()) :: {:ok, Enumerable.t()} | {:error, term()}

Returns all entries in the given tree.

tree_nth(tree, nth)

Specs

tree_nth(tree(), non_neg_integer()) ::
  {:ok, integer(), atom(), binary(), binary()} | {:error, term()}

Looks for a tree entry by its position in the given tree.

worktree_add(repo, name, path, ref)

Specs

worktree_add(repo(), binary(), binary(), binary() | :undefined) ::
  {:ok, worktree()} | {:error, term()}

Adds a new working tree for the given repo

worktree_prune(worktree)

Specs

worktree_prune(worktree()) :: :ok | {:error, term()}

Prunes a working tree.