GitRekt.GitAgent (git.limo v0.3.8)

High-level API for running Git commands on a repository.

This module provides an API to manipulate Git repositories. In contrast to GitRekt.Git, it offers an abstraction for serializing Git commands via message passing. By doing so it allows multiple processes to manipulate a single repository simultaneously (see “Thread safety” in GitRekt.Git module).

At it core GitRekt.GitAgent implements GenServer. Therefore it makes it easy to fit into a supervision tree and furthermore, in a distributed environment.

An other major benefit is support for caching command results. In their nature Git objects are immutable; that is, once they have been created and stored in the data store, they cannot be modified. This allows for very naive caching strategy implementations without having to deal with cache invalidation.

Example

Let's start by rewriting the example exposed in the GitRekt.Git module:

alias GitRekt.GitAgent

# load repository
{:ok, agent} = GitAgent.start_link("tmp/my-repo.git")

# fetch master branch
{:ok, branch} = GitAgent.branch(agent, "master")

# fetch commit pointed by master
{:ok, commit} = GitAgent.peel(agent, branch)

# fetch commit author & message
{:ok, author} = GitAgent.commit_author(agent, commit)
{:ok, message} = GitAgent.commit_message(agent, commit)

IO.puts "Last commit by #{author.name} <#{author.email}>:"
IO.puts message

This look very similar to the original example altought the API slightly differs.

You might have noticed that the first argument of each Git function is agent. In our example the agent is a PID referencing a dedicated process started with start_link/1.

Note that replacing start_link/1 with GitRekt.Git.repository_open/1 would print the exact same output. This works because agent/0 can be either a process id (PID) or a GitRekt.Git.repo/0.

Transactions

You can execute a serie of commands inside a transaction.

In the following example, we use transaction/2 to retrieve all informations for a given commit:

GitAgent.transaction(agent, fn agent ->
  with {:ok, author} <- GitAgent.commit_author(agent, commit),
      {:ok, committer} <- GitAgent.commit_committer(agent, commit),
      {:ok, message} <- GitAgent.commit_message(agent, commit),
      {:ok, parents} <- GitAgent.commit_parents(agent, commit),
      {:ok, timestamp} <- GitAgent.commit_timestamp(agent, commit),
      {:ok, gpg_sig} <- GitAgent.commit_gpg_signature(agent, commit) do
    {:ok, %{
      oid: commit.oid,
      author: author,
      committer: committer,
      message: message,
      parents: Enum.to_list(parents),
      timestamp: timestamp,
      gpg_sig: gpg_sig
    }}
  end
end)

Transactions provide a simple entry point for implementing more complex commands. The function is executed by the agent in a single request; avoiding the costs of making six consecutive GenServer.call/3.

You can also use transactions to cache the result of a serie of commands. Let's wrap our previous example in a function and give the transaction an unique identifier.

def commit_info(agent, commit) do
  # identifier: {:commit_info, oid}
  GitAgent.transaction(agent, {:commit_info, commit.oid}, fn agent -> ... end)
end

By using a unique identifier, we tell GitRekt.GitAgent that the result of our function should be cached.

Here's the log output for two consecutive commit_info/2 calls:

[debug] [Git Agent] transaction(:commit_info, "b662d32") executed in 361 µs
[debug] [Git Agent] > commit_author(<GitCommit:b662d32>) executed in 6 µs
[debug] [Git Agent] > commit_committer(<GitCommit:b662d32>) executed in 5 µs
[debug] [Git Agent] > commit_message(<GitCommit:b662d32>) executed in 1 µs
[debug] [Git Agent] > commit_parents(<GitCommit:b662d32>) executed in 4 µs
[debug] [Git Agent] > commit_timestamp(<GitCommit:b662d32>) executed in 11 µs
[debug] [Git Agent] > commit_gpg_signature(<GitCommit:b662d32>) executed in 6 µs
[debug] [Git Agent] transaction(:commit_info, "b662d32") executed in ⚡ 3 µs

We can observe that the first call executes the different commands one by one and cache the result while the second one fetches the result directly from the cache without having to actually run the transaction.

Note that transaction/3 can be called recursively and still benefit from caching.

Types

agent()

Specs

agent() :: pid() | GitRekt.Git.repo()

git_object()

Specs

git_object() ::
  GitRekt.GitCommit.t()
  | GitRekt.GitBlob.t()
  | GitRekt.GitTree.t()
  | GitRekt.GitTag.t()

git_revision()

Specs

git_revision() ::
  GitRekt.GitRef.t() | GitRekt.GitTag.t() | GitRekt.GitCommit.t()

Functions

blob_content(agent, blob, opts \\ [])

Specs

blob_content(agent(), GitRekt.GitBlob.t(), keyword()) ::
  {:ok, binary()} | {:error, term()}

Returns the content of the given blob.

blob_size(agent, blob, opts \\ [])

Specs

blob_size(agent(), GitRekt.GitBlob.t(), keyword()) ::
  {:ok, non_neg_integer()} | {:error, term()}

Returns the size in byte of the given blob.

branch(agent, name, opts \\ [])

Specs

branch(agent(), binary(), keyword()) ::
  {:ok, GitRekt.GitRef.t() | {GitRekt.GitRef.t(), GitRekt.GitCommit.t()}}
  | {:error, term()}

Returns the Git branch with the given name.

branches(agent, opts \\ [])

Specs

branches(agent(), keyword()) :: {:ok, Enumerable.t()} | {:error, term()}

Returns all Git branches.

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

commit_author(agent, commit, opts \\ [])

Specs

commit_author(agent(), GitRekt.GitCommit.t(), keyword()) ::
  {:ok, map()} | {:error, term()}

Returns the author of the given commit.

commit_committer(agent, commit, opts \\ [])

Specs

commit_committer(agent(), GitRekt.GitCommit.t(), keyword()) ::
  {:ok, map()} | {:error, term()}

Returns the committer of the given commit.

commit_create(agent, author, committer, message, tree_oid, parents_oids, opts \\ [])

Specs

commit_create(
  agent(),
  map(),
  map(),
  binary(),
  GitRekt.Git.oid(),
  [GitRekt.Git.oid()],
  keyword()
) :: {:ok, GitRekt.Git.oid()} | {:error, term()}

Creates a commit with the given tree_oid and parents_oid.

commit_gpg_signature(agent, commit, opts \\ [])

Specs

commit_gpg_signature(agent(), GitRekt.GitCommit.t(), keyword()) ::
  {:ok, binary()} | {:error, term()}

Returns the GPG signature of the given commit.

commit_message(agent, commit, opts \\ [])

Specs

commit_message(agent(), GitRekt.GitCommit.t(), keyword()) ::
  {:ok, binary()} | {:error, term()}

Returns the message of the given commit.

commit_parents(agent, commit, opts \\ [])

Specs

commit_parents(agent(), GitRekt.GitCommit.t(), keyword()) ::
  {:ok, Enumerable.t()} | {:error, term()}

Returns the parent of the given commit.

commit_timestamp(agent, commit, opts \\ [])

Specs

commit_timestamp(agent(), GitRekt.GitCommit.t(), keyword()) ::
  {:ok, DateTime.t()} | {:error, term()}

Returns the timestamp of the given commit.

diff(agent, obj1, obj2, opts \\ [])

Specs

diff(
  agent(),
  git_revision() | GitRekt.GitTree.t(),
  git_revision() | GitRekt.GitTree.t(),
  keyword()
) :: {:ok, GitRekt.GitDiff.t()} | {:error, term()}

Returns the Git diff of obj1 and obj2.

diff_deltas(agent, diff, opts \\ [])

Specs

diff_deltas(agent(), GitRekt.GitDiff.t(), keyword()) ::
  {:ok, map()} | {:error, term()}

Returns the deltas of the given diff.

diff_format(agent, diff, opts \\ [])

Specs

diff_format(agent(), GitRekt.GitDiff.t(), keyword()) ::
  {:ok, binary()} | {:error, term()}

Returns a binary formated representation of the given diff.

diff_stats(agent, diff, opts \\ [])

Specs

diff_stats(agent(), GitRekt.GitDiff.t(), keyword()) ::
  {:ok, map()} | {:error, term()}

Returns the stats of the given diff.

empty?(agent, opts \\ [])

Specs

empty?(agent(), keyword()) :: {:ok, boolean()} | {:error, term()}

Returns true if the repository is empty; otherwise returns false.

graph_ahead_behind(agent, local, upstream, opts \\ [])

Specs

graph_ahead_behind(agent(), GitRekt.Git.oid(), GitRekt.Git.oid(), keyword()) ::
  {:ok, {non_neg_integer(), non_neg_integer()}} | {:error, term()}

Returns the number of unique commits between two commit objects.

head(agent, opts \\ [])

Specs

head(agent(), keyword()) :: {:ok, GitRekt.GitRef.t()} | {:error, term()}

Returns the Git reference for HEAD.

history(agent, revision, opts \\ [])

Specs

history(agent(), git_revision(), keyword()) ::
  {:ok, Enumerable.t()} | {:error, term()}

Returns the Git commit history of the given revision.

index(agent, opts \\ [])

Specs

index(agent(), keyword()) :: {:ok, GitRekt.GitIndex.t()} | {:error, term()}

Returns the Git index of the repository.

index_add(agent, index, index_entry, opts \\ [])

Specs

index_add(agent(), GitRekt.GitIndex.t(), GitRekt.GitIndexEntry.t(), keyword()) ::
  :ok | {:error, term()}

Adds index_entry to the gieven index.

index_add(agent, index, oid, path, file_size, mode, opts \\ [])

Specs

index_add(
  agent(),
  GitRekt.GitIndex.t(),
  GitRekt.Git.oid(),
  Path.t(),
  non_neg_integer(),
  pos_integer(),
  keyword()
) :: :ok | {:error, term()}

Adds an index entry to the given index.

index_read_tree(agent, index, tree, opts \\ [])

Specs

index_read_tree(agent(), GitRekt.GitIndex.t(), GitRekt.GitTree.t(), keyword()) ::
  :ok | {:error, term()}

Reads the given tree into the given index.

index_remove(agent, index, path, opts \\ [])

Specs

index_remove(agent(), GitRekt.GitIndex.t(), Path.t(), keyword()) ::
  :ok | {:error, term()}

Removes an index entry from the given index.

index_remove_dir(agent, index, path, opts \\ [])

Specs

index_remove_dir(agent(), GitRekt.GitIndex.t(), Path.t(), keyword()) ::
  :ok | {:error, term()}

Remove all entries from the given index under a given directory path.

index_write_tree(agent, index, opts \\ [])

Specs

index_write_tree(agent(), GitRekt.GitIndex.t(), keyword()) ::
  {:ok, GitRekt.Git.oid()} | {:error, term()}

Writes the given index as a tree.

object(agent, oid, opts \\ [])

Specs

object(agent(), GitRekt.Git.oid(), keyword()) ::
  {:ok, git_object()} | {:error, term()}

Returns the Git object with the given oid.

odb(agent, opts \\ [])

Specs

odb(agent(), keyword()) :: {:ok, GitRekt.GitOdb.t()}

Returns the ODB.

odb_object_exists?(agent, odb, oid, opts \\ [])

Specs

odb_object_exists?(agent(), GitRekt.GitOdb.t(), GitRekt.Git.oid(), keyword()) ::
  {:ok, boolean()} | {:error, term()}

Returns true if the given oid exists in odb; elsewise returns false.

odb_read(agent, odb, oid, opts \\ [])

Specs

odb_read(agent(), GitRekt.GitOdb.t(), GitRekt.Git.oid(), keyword()) ::
  {:ok, {GitRekt.Git.obj_type(), binary()}} | {:error, term()}

Return the raw data of the odb object with the given oid.

odb_write(agent, odb, data, type, opts \\ [])

Specs

odb_write(agent(), GitRekt.GitOdb.t(), binary(), atom(), keyword()) ::
  {:ok, GitRekt.Git.oid()} | {:error, term()}

Writes the given data into the odb.

odb_writepack(agent, opts \\ [])

Specs

odb_writepack(agent(), keyword()) ::
  {:ok, GitRekt.GitWritePack.t()} | {:error, term()}

Returns an ODB writepack.

odb_writepack_append(agent, writepack, data, progress, opts \\ [])

Specs

odb_writepack_append(
  agent(),
  GitRekt.GitWritePack.t(),
  binary(),
  GitRekt.Git.odb_writepack_progress(),
  keyword()
) :: {:ok, GitRekt.Git.odb_writepack_progress()} | {:error, term()}

Appends data to the given writepack.

odb_writepack_commit(agent, writepack, progress, opts \\ [])

Specs

odb_writepack_commit(
  agent(),
  GitRekt.GitWritePack.t(),
  GitRekt.Git.odb_writepack_progress(),
  keyword()
) :: {:ok, GitRekt.Git.odb_writepack_progress()} | {:error, term()}

Commits the given writepack to the ODB.

pack_create(agent, oids, opts \\ [])

Specs

pack_create(agent(), [GitRekt.Git.oid()], keyword()) ::
  {:ok, binary()} | {:error, term()}

Returns a Git PACK representation of the given oids.

peel(agent, obj, opts \\ [])

Specs

peel(agent(), git_revision() | GitRekt.GitTreeEntry.t(), keyword()) ::
  {:ok, git_object()} | {:error, term()}

Peels the given obj until a Git object of the specified type is met.

reference(agent, name, opts \\ [])

Specs

reference(agent(), binary(), keyword()) ::
  {:ok, GitRekt.GitRef.t() | {GitRekt.GitRef.t(), GitRekt.GitCommit.t()}}
  | {:error, term()}

Returns the Git reference with the given name.

reference_create(agent, name, type, target, opts \\ [])

Specs

reference_create(
  agent(),
  binary(),
  atom(),
  GitRekt.Git.oid() | binary(),
  keyword()
) :: :ok | {:error, term()}

Creates a reference with the given name and target.

reference_delete(agent, name, opts \\ [])

Specs

reference_delete(agent(), binary(), keyword()) :: :ok | {:error, term()}

Deletes a reference with the given name.

references(agent, opts \\ [])

Specs

references(agent(), keyword()) :: {:ok, Enumerable.t()} | {:error, term()}

Returns all Git references matching the given glob.

revision(agent, spec, opts \\ [])

Specs

revision(agent(), binary(), keyword()) ::
  {:ok, {GitRekt.GitRef.t() | GitRekt.GitTag.t(), GitRekt.GitCommit.t() | nil}}
  | {:error, term()}

Returns the Git object matching the given spec.

start_link(path, opts \\ [])

Specs

start_link(Path.t(), keyword()) :: GenServer.on_start()

Starts a Git agent linked to the current process for the repository at the given path.

tag(agent, name, opts \\ [])

Specs

tag(agent(), binary(), keyword()) ::
  {:ok,
   GitRekt.GitRef.t()
   | GitRekt.GitTag.t()
   | {GitRekt.GitRef.t() | GitRekt.GitTag.t(), GitRekt.GitCommit.t()}}
  | {:error, term()}

Returns the Git tag with the given name.

tag_author(agent, tag, opts \\ [])

Specs

tag_author(agent(), GitRekt.GitTag.t(), keyword()) ::
  {:ok, map()} | {:error, term()}

Returns the Git tag author of the given tag.

tag_message(agent, tag, opts \\ [])

Specs

tag_message(agent(), GitRekt.GitTag.t(), keyword()) ::
  {:ok, binary()} | {:error, term()}

Returns the Git tag message of the given tag.

tags(agent, opts \\ [])

Specs

tags(agent(), keyword()) :: {:ok, Enumerable.t()} | {:error, term()}

Returns all Git tags.

transaction(agent, name \\ nil, cb, opts \\ [])

Specs

transaction(
  agent(),
  term(),
  (GitRekt.Git.repo() -> {:ok, term()} | {:error, term()}),
  keyword()
) :: {:ok, term()} | {:error, term()}

Executes the given cb inside a transaction.

tree(agent, revision, opts \\ [])

Specs

tree(agent(), git_revision(), keyword()) ::
  {:ok, GitRekt.GitTree.t()} | {:error, term()}

Returns the Git tree of the given revision.

tree_entries(agent, revision, opts \\ [])

Specs

tree_entries(agent(), git_revision() | GitRekt.GitTree.t(), keyword()) ::
  {:ok, Enumerable.t()} | {:error, term()}

Returns the Git tree entries of the given tree.

tree_entry_by_id(agent, revision, oid, opts \\ [])

Specs

tree_entry_by_id(
  agent(),
  git_revision() | GitRekt.GitTree.t(),
  GitRekt.Git.oid(),
  keyword()
) :: {:ok, GitRekt.GitTreeEntry.t()} | {:error, term()}

Returns the Git tree entry for the given revision and oid.

tree_entry_by_path(agent, revision, path, opts \\ [])

Specs

tree_entry_by_path(
  agent(),
  git_revision() | GitRekt.GitTree.t(),
  Path.t(),
  keyword()
) ::
  {:ok,
   GitRekt.GitTreeEntry.t() | {GitRekt.GitTreeEntry.t(), GitRekt.GitCommit.t()}}
  | {:error, term()}

Returns the Git tree entry for the given revision and path.