Git Gud v0.1.0 GitRekt.WireProtocol behaviour

Conveniences for Git transport protocol and server side commands.

Functions

decode(pkt)

decode(binary()) :: Stream.t()

Returns a stream of decoded PKT-LINEs for the given pkt.

done?(service)

done?(struct()) :: boolean()

Returns true if service is done; elsewhise returns false.

encode(lines)

encode(Enumerable.t()) :: iolist()

Returns an PKT-LINE encoded representation of the given lines.

new(repo, executable, init_values \\ [])

Returns a new service object for the given repo and executable.

next(service, data \\ :discovery)

next(struct(), binary() | :discovery) :: {struct(), iolist()}

Runs the given service to the next step.

pkt_line(data \\ :flush)

pkt_line(binary() | :flush) :: binary()

Returns the given data formatted as PKT-LINE

reference_discovery(repo, service)

reference_discovery(GitRekt.Git.repo(), binary()) :: iolist()

Returns a stream describing each ref and it current value.

run(service, data \\ :discovery, opts \\ [])

run(struct(), binary() | :discovery, keyword()) :: {struct(), iolist()}

Runs all the steps of the given service at once.

skip(service)

skip(struct()) :: struct()

Sets the given service to the next logical step without performing any action.

Callbacks

next(struct, list)

next(struct(), [term()]) :: {struct(), [term()]}

Callback used to transist a service to the next step.

skip(struct)

skip(struct()) :: struct()

Callback used to transist a service to the next step without performing any action.