Git Gud v0.1.0 GitGud.Email

Email schema and helper functions.

An GitGud.Email is used for a many different tasks such as user authentication & verification, email notifications, identification of Git commit authors, etc.

Every GitGud.User has at least one email address. In order to be taken in account, an email address must be verified first. See verify/1 for more details.

Once verified, an email address can be used to authenticate users (see GitGud.Auth.check_credentials/2) and resolve Git commit authors.

In order to associate Git commits to a specific GitGud.User account, every user can have has many email addresses as he likes. Once verified, emails appearing in Git commits will automatically be linked to the associated user. See GitGud.GPGKey for more details on how to verify GPG (or S/MIME) signed commits.

Types

t()

t() :: %GitGud.Email{
  __meta__: term(),
  address: binary(),
  id: pos_integer(),
  inserted_at: NaiveDateTime.t(),
  user: GitGud.User.t(),
  user_id: pos_integer(),
  verified: boolean(),
  verified_at: NaiveDateTime.t()
}

Functions

changeset(email, params \\ %{})

changeset(t(), map()) :: Ecto.Changeset.t()

Returns an email changeset for the given params.

create!(params)

create!(map() | keyword()) :: t()

Similar to create/1, but raises an Ecto.InvalidChangesetError if an error occurs.

create(params)

create(map() | keyword()) :: {:ok, t()} | {:error, Ecto.Changeset.t()}

Creates a new email with the given params.

{:ok, email} = GitGud.Email.create(user_id: user.id, address: "m.flach@almightycouch.com")

This function validates the given params using changeset/2.

delete!(email)

delete!(t()) :: t()

Similar to delete!/1, but raises an Ecto.InvalidChangesetError if an error occurs.

delete(email)

delete(t()) :: {:ok, t()} | {:error, Ecto.Changeset.t()}

Deletes the given email.

verify!(email)

verify!(t()) :: t()

Similar to verify/1, but raises an Ecto.InvalidChangesetError if an error occurs.

verify(email)

verify(t()) :: {:ok, t()} | {:error, Ecto.Changeset.t()}

Verifies the given email.