GitGud.GraphQL.Resolvers (git.limo v0.3.8)

Module providing resolution functions for GraphQL related queries.

Functions

close_issue(args, info)

Specs

close_issue(%{id: pos_integer()}, Absinthe.Resolution.t()) ::
  {:ok, GitGud.Issue.t()} | {:error, term()}

Closes a repository issue.

comment_author(comment, args, info)

Specs

comment_author(GitGud.Comment.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, GitGud.User.t()} | {:error, term()}

Resolves the author for a given comment.

comment_html(comment, args, info)

Specs

comment_html(GitGud.Comment.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, binary()} | {:error, term()}

Resolves the HTML content of a given comment.

comment_permissions(comment, args, info)

Specs

comment_permissions(GitGud.Comment.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, [atom()]} | {:error, term()}

Resolves the permissions for a given comment.

comment_repo(comment, args, info)

Specs

comment_repo(GitGud.Comment.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, GitGud.Repo.t()} | {:error, term()}

Resolves the repository for a given comment.

comment_revision_author(revision, args, info)

Specs

comment_revision_author(
  GitGud.CommentRevision.t(),
  %{},
  Absinthe.Resolution.t()
) :: {:ok, GitGud.User.t()} | {:error, term()}

Resolves the author for a given comment revision.

comment_revision_html(revision, args, info)

Specs

comment_revision_html(GitGud.CommentRevision.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, binary()} | {:error, term()}

Resolves the HTML content of a given comment revision.

comment_revisions(args, info)

Specs

comment_revisions(map(), Absinthe.Resolution.t()) ::
  {:ok, Absinthe.Relay.Connection.t()} | {:error, term()}

Resolves revisions for a comment.

comment_thread_type(result, info)

Specs

comment_thread_type(
  GitGud.Issue.t() | GitGud.CommitLineReview.t(),
  Absinthe.Resolution.t()
) :: atom()

Returns the search result type.

commit_line_review(commit, args, info)

Specs

commit_line_review(GitRekt.GitCommit.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, GitGud.CommitLineReview.t()} | {:error, term()}

Resolves the line review for a given Git commit object.

commit_line_review_comments(args, info)

Specs

commit_line_review_comments(map(), Absinthe.Resolution.t()) ::
  {:ok, Absinthe.Relay.Connection.t()} | {:error, term()}

Resolves comments for a commit line review.

commit_line_review_repo(commit_line_review, args, info)

Specs

commit_line_review_repo(
  GitGud.CommitLineReview.t(),
  %{},
  Absinthe.Resolution.t()
) :: {:ok, GitGud.Repo.t()} | {:error, term()}

Resolves the repository for a given commit_line_review.

commit_line_reviews(commit, args, info)

Specs

commit_line_reviews(GitRekt.GitCommit.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, [GitGud.CommitLineReview.t()]} | {:error, term()}

Resolves the line reviews for a given Git commit object.

create_commit_line_review_comment(args, info)

Specs

create_commit_line_review_comment(
  %{
    repo_id: pos_integer(),
    commit_oid: Git.oid(),
    blob_oid: Git.oid(),
    hunk: non_neg_integer(),
    line: non_neg_integer(),
    body: binary()
  },
  Absinthe.Resolution.t()
) :: {:ok, GitGud.Comment.t()} | {:error, term()}

Creates a Git line commit review comment.

create_issue_comment(args, info)

Specs

create_issue_comment(
  %{id: pos_integer(), body: binary()},
  Absinthe.Resolution.t()
) :: {:ok, GitGud.Comment.t()} | {:error, term()}

Creates a repository issue comment.

delete_comment(args, info)

Specs

delete_comment(%{id: pos_integer()}, Absinthe.Resolution.t()) ::
  {:ok, GitGud.Comment.t()} | {:error, term()}

Deletes a comment.

git_actor_type(actor, info)

Specs

git_actor_type(GitGud.User.t() | map(), Absinthe.Resolution.t()) :: atom()

Resolves the type for a given Git actor.

git_blob_size(blob, args, info)

Specs

git_blob_size(GitRekt.GitBlob.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, integer()} | {:error, term()}

Resolves the content length for a given Git blob object.

git_commit_author(commit, args, info)

Specs

git_commit_author(GitRekt.GitCommit.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, GitGud.User.t() | map()} | {:error, term()}

Resolves the author for a given Git commit object.

git_commit_committer(commit, args, info)

Specs

git_commit_committer(GitRekt.GitCommit.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, GitGud.User.t() | map()} | {:error, term()}

Resolves the committer for a given Git commit object.

git_commit_message(commit, args, info)

Specs

git_commit_message(GitRekt.GitCommit.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, binary()} | {:error, term()}

Resolves the message for a given Git commit object.

git_commit_parents(args, info)

Specs

git_commit_parents(map(), Absinthe.Resolution.t()) ::
  {:ok, Absinthe.Relay.Connection.t()} | {:error, term()}

Resolves the parents for a given Git commit object.

git_commit_timestamp(commit, args, info)

Specs

git_commit_timestamp(GitRekt.GitCommit.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, DateTime.t()} | {:error, term()}

Resolves the timestamp for a given Git commit object.

git_history(args, info)

Specs

git_history(map(), Absinthe.Resolution.t()) ::
  {:ok, Absinthe.Relay.Connection.t()} | {:error, term()}

Resolves the commit history starting from the given Git revision object.

git_object_type(object, info)

Specs

git_object_type(GitRekt.GitAgent.git_object(), Absinthe.Resolution.t()) ::
  atom()

Resolves the type for a given Git object.

git_reference_target(reference, args, info)

Specs

git_reference_target(GitReference.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, GitRekt.GitAgent.git_object()} | {:error, term()}

Resolves the Git target for the given Git reference object.

git_reference_type(reference, args, info)

Specs

git_reference_type(GitReference.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, atom()} | {:error, term()}

Resolves the type for a given Git tag.

git_tag_author(tag, args, info)

Specs

git_tag_author(GitRekt.GitTag.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, GitGud.User.t() | map()} | {:error, term()}

Resolves the author for a given Git tag object.

git_tag_message(tag, args, info)

Specs

git_tag_message(GitRekt.GitTag.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, binary()} | {:error, term()}

Resolves the message for a given Git tag object.

git_tag_target(tag, args, info)

Specs

git_tag_target(GitRekt.GitTag.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, GitRekt.GitAgent.git_object()} | {:error, term()}

Resolves the Git target for the given Git tag object.

git_tag_type(ref, info)

Specs

git_tag_type(GitReference.t() | GitRekt.GitTag.t(), Absinthe.Resolution.t()) ::
  {:ok, atom()} | {:error, term()}

Resolves the type for a given Git tag.

git_tree(revision, args, info)

Specs

git_tree(GitRekt.GitAgent.git_revision(), %{}, Absinthe.Resolution.t()) ::
  {:ok, GitRekt.GitTree.t()} | {:error, term()}

Resolves the tree for a given Git commit object.

git_tree_entries(args, info)

Specs

git_tree_entries(map(), Absinthe.Resolution.t()) ::
  {:ok, Absinthe.Relay.Connection.t()} | {:error, term()}

Resolves the tree entries for a given Git tree object.

git_tree_entry_target(tree_entry, args, info)

Specs

git_tree_entry_target(GitGud.Repo.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, GitRekt.GitTree.t() | GitRekt.GitBlob.t()} | {:error, term()}

Returns the underlying Git object for a given Git tree_entry object.

issue_author(issue, args, info)

Specs

issue_author(GitGud.Issue.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, GitGud.User.t()} | {:error, term()}

Resolves the author for a given comment.

issue_commit_reference_event_oid(event, args, info)

Specs

issue_commit_reference_event_oid(map(), %{}, Absinthe.Resolution.t()) ::
  {:ok, binary()}

Resolves the commit OID for a given issue event.

issue_commit_reference_event_url(map, args, info)

Specs

issue_commit_reference_event_url(map(), %{}, Absinthe.Resolution.t()) ::
  {:ok, binary()}

Resolves the commit URL for a given issue event.

issue_event_field(event, field, args, info)

Specs

issue_event_field(map(), binary(), %{}, Absinthe.Resolution.t()) :: {:ok, any()}

Resolves the field of a given issue event.

issue_event_timestamp(event, args, info)

Specs

issue_event_timestamp(map(), %{}, Absinthe.Resolution.t()) ::
  {:ok, NaiveDateTime.t()} | {:error, term()}

Resolves the timestamp for a given issue event.

issue_event_type(event, info)

Specs

issue_event_type(map(), Absinthe.Resolution.t()) :: atom()

Resolves the type of a given issue event.

issue_event_user(event, args, info)

Specs

issue_event_user(map(), %{}, Absinthe.Resolution.t()) ::
  {:ok, GitGud.User.t()} | {:error, term()}

Resolves the user for a given issue event.

issue_labels_update_event_pull_labels(event, args, info)

Specs

issue_labels_update_event_pull_labels(map(), %{}, Absinthe.Resolution.t()) ::
  {:ok, [binary()]}

Resolves the pull labels ids for a given issue event.

issue_labels_update_event_push_labels(event, args, info)

Specs

issue_labels_update_event_push_labels(map(), %{}, Absinthe.Resolution.t()) ::
  {:ok, [binary()]}

Resolves the push labels ids for a given issue event.

issue_permissions(issue, args, info)

Specs

issue_permissions(GitGud.Issue.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, [atom()]} | {:error, term()}

Resolves the permissions for a given issue.

issue_replies(args, info)

Specs

issue_replies(map(), Absinthe.Resolution.t()) ::
  {:ok, Absinthe.Relay.Connection.t()} | {:error, term()}

Resolves replies for an issue.

issue_repo(issue, args, info)

Specs

issue_repo(GitGud.Issue.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, GitGud.Repo.t()} | {:error, term()}

Resolves the repository for a given issue.

node(node_type, info, opts \\ [])

Specs

node(map(), Absinthe.Resolution.t(), keyword()) ::
  {:ok, map()} | {:error, term()}

Resolves a node object.

node_type(node, info)

Specs

node_type(map(), Absinthe.Resolution.t()) :: atom() | nil

Resolves a node object type.

reopen_issue(args, info)

Specs

reopen_issue(%{id: pos_integer()}, Absinthe.Resolution.t()) ::
  {:ok, GitGud.Issue.t()} | {:error, term()}

Reopens a repository issue.

repo(args, info)

Specs

repo(map(), Absinthe.Resolution.t()) ::
  {:ok, GitGud.Repo.t()} | {:error, term()}

Resolves a repository object.

repo_head(repo, args, info)

Specs

repo_head(GitGud.Repo.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, GitReference.t()} | {:error, term()}

Resolves the default branch object for a given repo.

repo_issue(repo, args, info)

Specs

repo_issue(GitGud.Repo.t(), %{number: number()}, Absinthe.Resolution.t()) ::
  {:ok, integer()} | {:error, term()}

Resolves a single issue of a given repo.

repo_issue_labels(repo, args, info)

Specs

repo_issue_labels(GitGud.Repo.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, [GitGud.IssueLabel.t()]}

Resolvers the issue labels associated to a given repo.

repo_issues(args, info)

Specs

repo_issues(map(), Absinthe.Resolution.t()) ::
  {:ok, Absinthe.Relay.Connection.t()} | {:error, term()}

Resolves issues for a repository.

repo_object(repo, args, info)

Specs

repo_object(GitGud.Repo.t(), %{oid: binary()}, Absinthe.Resolution.t()) ::
  {:ok, GitRekt.GitAgent.git_object()} | {:error, term()}

Resolves a Git object by OID for a given repo.

repo_ref(repo, args, info)

Specs

repo_ref(GitGud.Repo.t(), %{name: binary()}, Absinthe.Resolution.t()) ::
  {:ok, GitReference.t()} | {:error, term()}

Resolves a Git reference object by name for a given repo.

repo_refs(args, info)

Specs

repo_refs(map(), Absinthe.Resolution.t()) ::
  {:ok, Absinthe.Relay.Connection.t()} | {:error, term()}

Resolves all Git reference objects for a given repo.

repo_revision(repo, args, info)

Specs

repo_revision(GitGud.Repo.t(), %{spec: binary()}, Absinthe.Resolution.t()) ::
  {:ok, GitRekt.GitAgent.git_object()} | {:error, term()}

Resolves a Git commit by revision spec for a given repo.

repo_tag(repo, args, info)

Specs

repo_tag(GitGud.Repo.t(), %{name: binary()}, Absinthe.Resolution.t()) ::
  {:ok, GitReference.t() | GitRekt.GitTag.t()} | {:error, term()}

Resolves a Git tag object by name for a given repo.

repo_tags(args, info)

Specs

repo_tags(map(), Absinthe.Resolution.t()) ::
  {:ok, Absinthe.Relay.Connection.t()} | {:error, term()}

Resolves all Git tag objects for a given repo.

search(args, info)

Specs

search(map(), Absinthe.Resolution.t()) ::
  {:ok, Absinthe.Relay.Connection.t()} | {:error, term()}

Resolves a list of users for a given search term.

search_result_type(result, info)

Specs

search_result_type(GitGud.User.t() | GitGud.Repo.t(), Absinthe.Resolution.t()) ::
  atom()

Returns the search result type.

update_comment(args, info)

Specs

update_comment(%{id: pos_integer(), body: binary()}, Absinthe.Resolution.t()) ::
  {:ok, GitGud.Comment.t()} | {:error, term()}

Updates a comment.

update_issue_labels(args, info)

Specs

update_issue_labels(
  %{id: pos_integer(), label_id: pos_integer()},
  Absinthe.Resolution.t()
) :: {:ok, GitGud.Issue.t()} | {:error, term()}

Updates the label of an issue.

update_issue_title(args, info)

Specs

update_issue_title(
  %{id: pos_integer(), title: binary()},
  Absinthe.Resolution.t()
) :: {:ok, GitGud.Issue.t()} | {:error, term()}

Updates the title of an issue.

url(resource, args, info)

Specs

url(map(), %{}, Absinthe.Resolution.t()) :: {:ok, binary()} | {:error, term()}

Resolves the URL of the given resource.

user(args, info)

Specs

user(%{login: binary()}, Absinthe.Resolution.t()) ::
  {:ok, GitGud.User.t()} | {:error, term()}

Resolves an user object by login.

user_public_email(user, args, info)

Specs

user_public_email(GitGud.User.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, binary()} | {:error, term()}

Resolves the public email for a given user.

user_repo(user, args, info)

Specs

user_repo(GitGud.User.t(), %{name: binary()}, Absinthe.Resolution.t()) ::
  {:ok, GitGud.Repo.t()} | {:error, term()}

Resolves a repository object by name for a given user.

user_repos(args, info)

Specs

user_repos(map(), Absinthe.Resolution.t()) ::
  {:ok, Absinthe.Relay.Connection.t()} | {:error, term()}

Resolves all repositories for an user.

viewer(args, info)

Specs

viewer(%{}, Absinthe.Resolution.t()) ::
  {:ok, GitGud.User.t()} | {:error, term()}

Resolves the authenticated user.