Git Gud v0.1.0 GitGud.GraphQL.Schema

GraphQL schema definition.

Functions

from_relay_id(global_id)

from_relay_id(Absinthe.Relay.Node.global_id()) :: pos_integer() | nil

Returns the source id for the given Relay global_id.

to_relay_id(node)

to_relay_id(struct()) :: Absinthe.Relay.Node.global_id() | nil

Returns the Relay global id for the given node.

to_relay_id(node_type, source_id)

to_relay_id(atom() | binary(), pos_integer()) ::
  Absinthe.Relay.Node.global_id() | nil

Returns the Relay global id for the given source_id.