Git Gud v0.1.0 GitGud.Maintainer

Repository maintainer schema and helper functions.

A GitGud.Maintainer is primarly used to associate GitGud.User to GitGud.Repo.

Each repository maintainer also has a permission defining which actions he is able to perform on the repository. Following permissions are available:

:read — can read and clone the repository.

:write — can read, clone and push to the repository.

:admin — can read, clone, push and administrate the repository.

By default, a newly created repository maintainer has the :read permission. Use update_permission/2 to change a maintainer’s permission.

Types

t()

t() :: %GitGud.Maintainer{
  __meta__: term(),
  id: term(),
  inserted_at: NaiveDateTime.t(),
  permission: binary(),
  repo: GitGud.Repo.t(),
  repo_id: pos_integer(),
  updated_at: NaiveDateTime.t(),
  user: GitGud.User.t(),
  user_id: pos_integer()
}

Functions

changeset(maintainer, params \\ %{})

changeset(t(), map()) :: Ecto.Changeset.t()

Returns a maintainer changeset for the given params.

create!(params)

create!(map() | keyword()) :: t()

Similar to create/1, but raises an Ecto.InvalidChangesetError if an error occurs.

create(params)

create(map() | keyword()) :: {:ok, t()} | {:error, Ecto.Changeset.t()}

Creates a new maintainer with the given params.

{:ok, maintainer} = GitGud.Maintainer.create(user_id: user.id, repo_id: repo.id)

This function validates the given params using changeset/2.

delete!(maintainer)

Similar to delete!/1, but raises an Ecto.InvalidChangesetError if an error occurs.

delete(maintainer)

delete(t()) :: {:ok, t()} | {:error, Ecto.Changeset.t()}

delete(t()) :: t()

Deletes the given maintainer.

update_permission!(maintainer, permission)

update_permission!(t(), binary()) :: t()

Similar to update_permission/2, but raises an Ecto.InvalidChangesetError if an error occurs.

update_permission!(repo, user, permission)

update_permission!(GitGud.Repo.t(), GitGud.User.t(), binary()) :: t()

Similar to update_permission/3, but raises an Ecto.NoResultsError if an error occurs.

update_permission(maintainer, permission)

update_permission(t(), binary()) ::
  {:ok, t()} | {:error, Ecto.Changeset.t()}

Updates the permission of the given maintainer.

update_permission(repo, user, permission)

update_permission(GitGud.Repo.t(), GitGud.User.t(), binary()) ::
  {:ok, t()} | :error

Updates the permission of the given user for the given repo.