Git Gud v0.1.0 GitGud.RepoQuery

Conveniences for GitGud.Repo related queries.

Types

user_param()

user_param() :: GitGud.User.t() | binary() | pos_integer()

Functions

by_id(id, opts \\ [])

by_id(pos_integer(), keyword()) :: GitGud.Repo.t() | nil

by_id([pos_integer()], keyword()) :: [GitGud.Repo.t()]

Returns a repository for the given id.

by_path(path, opts \\ [])

by_path(Path.t(), keyword()) :: GitGud.Repo.t() | nil

Returns a repository for the given path.

repo_query(id)

repo_query(pos_integer()) :: Ecto.Query.t()

repo_query([pos_integer()]) :: Ecto.Query.t()

Returns a query for fetching a repositories by id.

search(input, opts \\ [])

search(binary(), keyword()) :: [GitGud.User.t()]

Returns a list of users matching the given input.

search_query(input)

search_query(binary()) :: Ecto.Query.t()

Returns a query for searching repositories.

user_repo(user, name, opts \\ [])

user_repo(user_param(), binary(), keyword()) :: GitGud.Repo.t() | nil

Returns a single repository for the given user and name.

user_repo_query(user, name)

user_repo_query(user_param(), binary()) :: Ecto.Query.t()

Returns a query for fetching a single repository for the given user and name.

user_repos(user, opts \\ [])

user_repos(user_param(), keyword()) :: [GitGud.Repo.t()]

user_repos([user_param()], keyword()) :: [GitGud.Repo.t()]

Returns a list of repositories for the given user.

user_repos_query(user)

user_repos_query(user_param()) :: Ecto.Query.t()

user_repos_query([user_param()]) :: Ecto.Query.t()

Returns a query for fetching user repositories.