Git Gud v0.1.0 GitGud.SSHKey

Secure Shell (SSH) authentication key schema and helper functions.

An GitGud.SSHKey is used to authenticate users via Public Key Authentication in GitGud.SSHServer.

Types

t()

t() :: %GitGud.SSHKey{
  __meta__: term(),
  data: binary(),
  fingerprint: binary(),
  id: pos_integer(),
  inserted_at: NaiveDateTime.t(),
  last_used_at: NaiveDateTime.t(),
  name: binary(),
  user: GitGud.User.t(),
  user_id: pos_integer()
}

Functions

changeset(ssh_key, params \\ %{})

changeset(t(), map()) :: Ecto.Changeset.t()

Returns a SSH key changeset for the given params.

If :name is omitted and the given public key contains a comment, the comment will be used as :name.

This function computes (see :public_key.ssh_decode/2 and :public_key.ssh_hostkey_fingerprint/1) and stores a fingerprint of the public key. It does not save the given :data into the database.

create!(params)

create!(map() | keyword()) :: t()

Similar to create/1, but raises an Ecto.InvalidChangesetError if an error occurs.

create(params)

create(map() | keyword()) :: {:ok, t()} | {:error, Ecto.Changeset.t()}

Creates a new SSH key with the given params.

{:ok, ssh_key} = GitGud.SSHKey.create(user_id: user.id, name: "My SSH Key", data: "ssh-rsa AAAAB3NzaC1yc2...")

This function validates the given params using changeset/2.

delete!(ssh_key)

delete!(t()) :: t()

Similar to delete!/1, but raises an Ecto.InvalidChangesetError if an error occurs.

delete(ssh_key)

delete(t()) :: {:ok, t()} | {:error, Ecto.Changeset.t()}

Deletes the given ssh_key.

update_timestamp!(ssh_key)

update_timestamp!(t()) :: t()

Similar to update_timestamp/1, but raises an Ecto.InvalidChangesetError if an error occurs.

update_timestamp(ssh_key)

update_timestamp(t()) :: {:ok, t()} | {:error, Ecto.Changeset.t()}

Updates the :last_used_at timestamp for the the given ssh_key.