Git Gud v0.1.0 GitGud.SSHServer

Secure Shell (SSH) server providing support for Git server commands.

The server handles following Git commands:

git-receive-pack - corresponding server-side command to git push.

git-upload-pack - corresponding server-side command to git fetch.

Authentication

A registered GitGud.User can authenticate with following methods:

public-key - if any of the associated GitGud.SSHKey matches.

password - if the given credentials are correct.

interactive - interactive login prompt allowing several tries.

To clone a repository, run following command:

git clone 'ssh://redrabbit@localhost:8989/USER/REPO'

Authorization

In order to read and/or write to a repository, a user needs to have the required permissions.

See GitGud.Authorization for more details.

Types

t()

t() :: %GitGud.SSHServer{
  chan: :ssh_connection.ssh_channel_id(),
  conn: :ssh_connection.ssh_connection_ref(),
  repo: term(),
  service: Module.t(),
  user: GitGud.User.t()
}

Functions

child_spec(args)

child_spec([]) :: Supervisor.Spec.spec()

Returns a child-spec to use as part of a supervision tree.