Git Gud v0.1.0 GitGud.SmartHTTPBackend

Plug providing support for Git server commands over HTTP.

This plug handles following Git commands:

git-receive-pack - corresponding server-side command to git push.

git-upload-pack - corresponding server-side command to git fetch.

Authentication

A registered GitGud.User can authenticate over HTTP via Basic Authentication. This is only, required for commands requiring specific permissions (such as pushing commits and cloning private repos).

To clone a repository, run following command:

git clone 'http://localhost:4000/USER/REPO'

Authorization

In order to read and/or write to a repository, a user needs to have the required permissions.

See GitGud.Authorization for more details.

Functions

call(conn, opts)

Callback implementation for Plug.call/2.

init(opts)

Callback implementation for Plug.init/1.