GitGud.UserQuery (git.limo v0.3.8)

Conveniences for user related queries.

Functions

by_email(email, opts \\ [])

Specs

by_email(binary(), keyword()) :: GitGud.User.t() | nil

by_email([binary()], keyword()) :: [GitGud.User.t()]

Returns a user for the given email.

by_id(id, opts \\ [])

Specs

by_id(pos_integer(), keyword()) :: GitGud.User.t() | nil

by_id([pos_integer()], keyword()) :: [GitGud.User.t()]

Returns a user for the given id.

by_login(login, opts \\ [])

Specs

by_login(binary(), keyword()) :: GitGud.User.t() | nil

by_login([binary()], keyword()) :: [GitGud.User.t()]

Returns a user for the given login.

by_oauth(provider, id, opts \\ [])

Specs

by_oauth(binary(), binary(), keyword()) :: GitGud.User.t() | nil

Returns a user for the given authentication provider and id.

by_ssh_key(fingerprint, opts \\ [])

Specs

by_ssh_key(binary(), keyword()) :: GitGud.User.t() | nil

Returns a user for the given SSH key fingerprint.

search(input, opts \\ [])

Specs

search(binary(), keyword()) :: [GitGud.User.t()]

Returns a list of users matching the given input.