GitGud.Web.AuthenticationPlug (git.limo v0.3.8)

Plug providing support for multiple authentication methods.

Functions

authenticate(conn, opts)

Specs

authenticate(Plug.Conn.t(), keyword()) :: Plug.Conn.t()

Plug to authenticate conn with either authorization or session tokens.

authenticate_bearer_token(conn, opts)

Specs

authenticate_bearer_token(Plug.Conn.t(), keyword()) :: Plug.Conn.t()

Plug to authenticate conn with authorization tokens.

authenticate_session(conn, opts)

Specs

authenticate_session(Plug.Conn.t(), keyword()) :: Plug.Conn.t()

Plug to authenticate conn with session tokens.

authenticated?(conn)

Specs

authenticated?(Plug.Conn.t()) :: boolean()

Returns true if the given conn is authenticated; otherwise returns false.

authentication_token(conn)

Specs

authentication_token(Plug.Conn.t()) :: binary() | nil

Generates an authentication token.

authorized?(conn, resource, action)

Specs

authorized?(Plug.Conn.t(), any(), atom()) :: boolean()

Returns true if the given conn is allowed to perform action on resource; otherwhise returns false.

current_user(conn)

Specs

current_user(Plug.Conn.t()) :: GitGud.User.t() | nil

Returns the current user if conn is authenticated.

ensure_authenticated(conn, opts)

Specs

ensure_authenticated(Plug.Conn.t(), keyword()) :: Plug.Conn.t()

Plug to ensure that the request is authenticated.

If the given conn is not authenticated?/1, this prevents further plugs downstream from being invoked and returns a 401 Unauthenticated error.

verified?(conn)

Specs

verified?(Plug.Conn.t()) :: boolean()

Returns true if the given conn is authenticated with a verified user; otherwise returns false.