Git Gud v0.1.0 GitGud.Web.NavigationHelpers

Conveniences for routing and navigation.

Functions

current_route(conn)

current_route(Plug.Conn.t()) :: {atom(), atom()}

Returns the conn controller and action as tuple.

current_route?(conn, helper, action \\ [])

current_route?(Plug.Conn.t(), atom(), []) :: boolean()

current_route?(Plug.Conn.t(), atom(), [{:only, [atom()]}]) :: boolean()

current_route?(Plug.Conn.t(), atom(), [{:except, [atom()]}]) :: boolean()

current_route?(Plug.Conn.t(), atom(), atom()) :: boolean()

Returns true if conn matches the given route helper; otherwhise return false.

navigation_item(conn, helper, action \\ [], tag \\ :li, attrs \\ [], list)

navigation_item(Plug.Conn.t(), atom(), keyword() | atom(), atom(), keyword(), [
  {:do, term()}
]) :: binary()

Renders a navigation item for the given helper and action.