Git Gud v0.1.0 GitRekt.Git

Erlang NIF that exposes a subset of libgit2’s library functions.

Types

blob()

blob() :: reference()

commit()

commit() :: reference()

config()

config() :: reference()

diff()

diff() :: reference()

diff_delta()

diff_delta() :: {diff_file(), diff_file(), non_neg_integer(), non_neg_integer()}

diff_file()

diff_file() :: {oid(), binary(), integer(), non_neg_integer()}

diff_format()

diff_format() :: :patch | :patch_header | :raw | :name_only | :name_status

diff_hunk()

diff_hunk() :: {binary(), integer(), integer(), integer(), integer()}

diff_line()

diff_line() :: {char(), integer(), integer(), integer(), integer(), binary()}

index()

index() :: reference()

index_entry()

index_entry() ::
  {integer(), integer(), non_neg_integer(), non_neg_integer(),
   non_neg_integer(), non_neg_integer(), non_neg_integer(), integer(), binary(),
   non_neg_integer(), non_neg_integer(), binary()}

obj()

obj() :: blob() | commit() | tree() | tag()

obj_type()

obj_type() :: :blob | :commit | :tree | :tag

odb()

odb() :: reference()

odb_type()

odb_type() :: atom()

oid()

oid() :: binary()

pack()

pack() :: reference()

ref_iter()

ref_iter() :: reference()

ref_type()

ref_type() :: :oid | :symbolic

reflog_entry()

reflog_entry() ::
  {binary(), binary(), non_neg_integer(), non_neg_integer(), oid(), oid(),
   binary()}

repo()

repo() :: reference()

revwalk()

revwalk() :: reference()

revwalk_sort()

revwalk_sort() :: :topsort | :timesort | :reversesort

signature()

signature() :: {binary(), binary(), non_neg_integer(), non_neg_integer()}

tag()

tag() :: reference()

tree()

tree() :: reference()

tree_entry()

tree_entry() :: {integer(), :blob | :tree, oid(), binary()}

Functions

blob_content(blob)

blob_content(blob()) :: {:ok, binary()} | {:error, term()}

Returns the raw content of the given blob.

blob_size(blob)

blob_size(blob()) :: {:ok, integer()} | {:error, term()}

Returns the size in bytes of the given blob.

commit_author(commit)

commit_author(commit()) ::
  {:ok, binary(), binary(), non_neg_integer(), non_neg_integer()}
  | {:error, term()}

Returns the author of the given commit.

commit_create(repo, ref, author, commiter, encoding, message, tree, parents)

commit_create(
  repo(),
  binary() | :undefined,
  signature(),
  signature(),
  binary() | :undefined,
  binary(),
  oid(),
  [binary()]
) :: {:ok, oid()} | {:error, term()}

Creates a new commit with the given params.

commit_message(commit)

commit_message(commit()) :: {:ok, binary()} | {:error, term()}

Returns the message for the given commit.

commit_parent(commit, index)

commit_parent(commit(), non_neg_integer()) ::
  {:ok, oid(), commit()} | {:error, term()}

Looks for a parent commit of the given commit by its index.

commit_parent_count(commit)

commit_parent_count(commit()) :: oid()

Returns the number of parents for the given commit.

commit_parents(commit)

commit_parents(commit()) :: {:ok, [{oid(), commit()}]} | {:error, term()}

Returns parent commits of the given commit.

commit_time(commit)

commit_time(commit()) ::
  {:ok, non_neg_integer(), non_neg_integer()} | {:error, term()}

Returns the time of the given commit.

commit_tree(commit)

commit_tree(commit()) :: {:ok, oid(), tree()} | {:error, term()}

Returns the tree for the given commit.

commit_tree_id(commit)

commit_tree_id(commit()) :: oid()

Returns the tree id for the given commit.

config_get_bool(config, name)

config_get_bool(config(), binary()) :: {:ok, boolean()} | {:error, term()}

Returns the value of the config entry with the given name.

config_get_string(config, name)

config_get_string(config(), binary()) :: {:ok, binary()} | {:error, term()}

Returns the value of the config entry with the given name.

config_open(path)

config_open(binary()) :: {:ok, config()} | {:error, term()}

Returns a config handle for the given path.

config_set_bool(config, name, val)

config_set_bool(config(), binary(), boolean()) :: :ok | {:error, term()}

Sets the config entry with the given name to val.

config_set_string(config, name, val)

config_set_string(config(), binary(), binary()) :: :ok | {:error, term()}

Sets the config entry with the given name to val.

diff_delta_count(diff)

diff_delta_count(diff()) :: {:ok, non_neg_integer()} | {:error, term()}

Returns the number of deltas in the given diff.

diff_deltas(diff)

diff_deltas(diff()) ::
  {:ok, [{diff_delta(), [{diff_hunk(), [diff_line()]}]}]} | {:error, term()}

Returns a list of deltas for the given diff.

diff_format(diff, format \\ :patch)

diff_format(diff(), diff_format()) :: {:ok, binary()} | {:error, term()}

Returns a binary represention of the given diff.

diff_stats(diff)

diff_stats(diff()) ::
  {:ok, non_neg_integer(), non_neg_integer(), non_neg_integer()}
  | {:error, term()}

Returns stats for the given diff.

diff_tree(repo, old_tree, new_tree, opts \\ [])

Returns a diff with the difference between two tree objects.

enumerate(stream)

enumerate(Stream.t()) :: {:ok, Stream.t()} | {:error, term()}

Enumerates the given stream by calling Enum.to_list/1 on the internal enum.

index_add(index, entry)

index_add(index(), index_entry()) :: :ok | {:error, term()}

Adds or updates the given entry.

index_bypath(index, path, stage)

index_bypath(index(), Path.t(), non_neg_integer()) ::
  {:ok, integer(), integer(), non_neg_integer(), non_neg_integer(),
   non_neg_integer(), non_neg_integer(), non_neg_integer(), integer(), binary(),
   non_neg_integer(), non_neg_integer(), binary()}
  | {:error, term()}

Retrieves an entry contained in the index given its relative path.

index_clear(index)

index_clear(index()) :: :ok | {:error, term()}

Clears the contents (all the entries) of the given index.

index_count(index)

index_count(index()) :: non_neg_integer()

Returns the number of entries in the given index.

index_new()

index_new() :: {:ok, index()} | {:error, term()}

Creates an new in-memory index object.

index_nth(index, nth)

index_nth(index(), non_neg_integer()) ::
  {:ok, integer(), integer(), non_neg_integer(), non_neg_integer(),
   non_neg_integer(), non_neg_integer(), non_neg_integer(), integer(), binary(),
   non_neg_integer(), non_neg_integer(), binary()}
  | {:error, term()}

Looks for an entry by its position in the given index.

index_read_tree(index, tree)

index_read_tree(index(), tree()) :: :ok | {:error, term()}

Reads the given tree into the given index file with stats.

index_write(index)

index_write(index()) :: :ok | {:error, term()}

Writes the given index from memory back to disk using an atomic file lock.

index_write_tree(index)

index_write_tree(index()) :: {:ok, oid()} | {:error, term()}

Writes the given index as a tree.

index_write_tree(index, repo)

index_write_tree(index(), repo()) :: {:ok, oid()} | {:error, term()}

Writes the given index as a tree.

library_version()

library_version() :: {integer(), integer(), integer()}

Returns the libgit2 library version.

object_id(obj)

object_id(obj()) :: {:ok, oid()} | {:error, term()}

Returns the OID for the given obj.

object_lookup(repo, oid)

object_lookup(repo(), oid()) :: {:ok, obj_type(), obj()} | {:error, term()}

Looks for an object with the given oid.

object_repository(obj)

object_repository(obj()) :: {:ok, repo()} | {:error, term()}

Returns the repository that owns the given obj.

object_zlib_inflate(data)

object_zlib_inflate(binary()) ::
  {:ok, iodata(), non_neg_integer()} | {:error, term()}

Inflates the given data with zlib.

odb_object_exists?(odb, oid)

odb_object_exists?(odb(), oid()) :: boolean()

Returns true if the given oid exists in odb; elsewhise returns false.

odb_read(odb, oid)

odb_read(odb(), oid()) :: {:ok, obj_type(), binary()}

Return the uncompressed, raw data of an ODB object.

odb_write(odb, data, type)

odb_write(odb(), binary(), odb_type()) :: {:ok, oid()} | {:error, term()}

Writes the given data into the odb.

oid_fmt(oid)

oid_fmt(oid()) :: binary()

Returns the SHA hash for the given oid.

oid_fmt_short(oid)

oid_fmt_short(oid()) :: binary()

Returns the abbreviated SHA hash for the given oid.

oid_parse(hash)

oid_parse(binary()) :: oid()

Returns the OID for the given SHA hash.

pack_data(pack)

pack_data(pack()) :: {:ok, binary()} | {:error, term()}

Returns a PACK file for the given pack.

pack_insert_commit(pack, oid)

pack_insert_commit(pack(), oid()) :: :ok | {:error, term()}

Inserts commit as well as the completed referenced tree.

pack_insert_walk(pack, walk)

pack_insert_walk(pack(), revwalk()) :: :ok | {:error, term()}

Inserts objects as given by the walk.

pack_new(repo)

pack_new(repo()) :: {:ok, pack()} | {:error, term()}

Creates a new PACK object for the given repo.

pathspec_match_tree(tree, pathspec)

pathspec_match_tree(tree(), [binary()]) :: {:ok, boolean()} | {:error, term()}

Returns true if tree matches the given pathspec; otherwise returns false.

reference_create(repo, name, type, target, force \\ false)

reference_create(repo(), binary(), ref_type(), binary() | oid(), boolean()) ::
  :ok | {:error, term()}

Creates a new reference name which points to an object or to an other reference.

reference_delete(repo, name)

reference_delete(repo(), binary()) :: :ok | {:error, term()}

Deletes an existing reference.

reference_dwim(repo, short_name)

reference_dwim(repo(), binary()) ::
  {:ok, binary(), ref_type(), binary()} | {:error, term()}

Looks for a reference by DWIMing its short_name.

reference_glob(repo, glob)

reference_glob(repo(), binary()) :: {:ok, [binary()]} | {:error, term()}

Similar to reference_list/1 but allows glob patterns.

reference_iterator(repo, glob \\ :undefined)

reference_iterator(repo(), binary() | :undefined) ::
  {:ok, ref_iter()} | {:error, term()}

Returns an iterator for the references that match the specific glob pattern.

reference_list(repo)

reference_list(repo()) :: {:ok, [binary()]} | {:error, term()}

Returns all references for the given repo.

reference_log?(repo, name)

reference_log?(repo(), binary()) :: {:ok, boolean()} | {:error, term()}

Returns true if a reflog exists for the given reference name.

reference_lookup(repo, name)

reference_lookup(repo(), binary()) ::
  {:ok, binary(), ref_type(), binary()} | {:error, term()}

Looks for a reference by name.

reference_next(iter)

reference_next(ref_iter()) ::
  {:ok, binary(), binary(), ref_type(), binary()} | {:error, term()}

Returns the next reference.

reference_peel(repo, name, type \\ :undefined)

reference_peel(repo(), binary(), obj_type() | :undefined) ::
  {:ok, obj_type(), oid(), obj()} | {:error, term()}

Recursively peels the given reference name until an object of type type is found.

reference_resolve(repo, name)

reference_resolve(repo(), binary()) ::
  {:ok, binary(), binary(), oid()} | {:error, term()}

Resolves a symbolic reference to a direct reference.

reference_stream(repo, glob \\ :undefined)

reference_stream(repo(), binary() | :undefined) ::
  {:ok, Stream.t()} | {:error, term()}

Returns a stream for the references that match the specific glob pattern.

reference_to_id(repo, name)

reference_to_id(repo(), binary()) :: {:ok, oid()} | {:error, term()}

Looks for a reference by name and returns its id.

reflog_count(repo, name)

reflog_count(repo(), binary()) :: {:ok, pos_integer()} | {:error, term()}

Reads the number of entry for the given reflog name.

reflog_delete(repo, name)

reflog_delete(repo(), binary()) :: :ok | {:error, term()}

Deletes the reflog for the given reference name.

reflog_read(repo, name)

reflog_read(repo(), binary()) :: {:ok, [reflog_entry()]} | {:error, term()}

Reads the reflog for the given reference name.

repository_bare?(repo)

repository_bare?(repo()) :: boolean()

Returns true if repo is bare; elsewhise returns false.

repository_discover(path)

repository_discover(Path.t()) :: {:ok, Path.t()} | {:error, term()}

Looks for a repository and returns its path.

repository_empty?(repo)

repository_empty?(repo()) :: boolean()

Returns true if repo is empty; elsewhise returns false.

repository_get_config(repo)

repository_get_config(repo()) :: {:ok, config()} | {:error, term()}

Returns the config for the given repo.

repository_get_odb(repo)

repository_get_odb(repo()) :: {:ok, odb()} | {:error, term()}

Returns the ODB for the given repository.

repository_get_path(repo)

repository_get_path(repo()) :: Path.t()

Returns the absolute path for the given repo.

repository_get_workdir(repo)

repository_get_workdir(repo()) :: Path.t()

Returns the normalized path to the working directory for the given repo.

repository_init(path, bare \\ false)

repository_init(Path.t(), boolean()) :: {:ok, repo()} | {:error, term()}

Initializes a new repository at the given path.

repository_open(path)

repository_open(Path.t()) :: {:ok, repo()} | {:error, term()}

Returns a repository handle for the path.

revparse_ext(repo, revision)

revparse_ext(repo(), binary()) ::
  {:ok, obj(), obj_type(), oid(), binary() | nil} | {:error, term()}

Finds a single object and intermediate reference, as specified by the given revision.

revparse_single(repo, revision)

revparse_single(repo(), binary()) ::
  {:ok, obj(), obj_type(), oid()} | {:error, term()}

Finds a single object, as specified by the given revision.

revwalk_new(repo)

revwalk_new(repo()) :: {:ok, reference()} | {:error, term()}

Creates a new revision walk object for the given repo.

revwalk_next(walk)

revwalk_next(revwalk()) :: {:ok, oid()} | {:error, term()}

Returns the next commit from the given revision walk.

revwalk_pack(walk)

Returns a PACK file for the given walk.

revwalk_push(walk, oid, hide \\ false)

revwalk_push(revwalk(), oid(), boolean()) :: :ok | {:error, term()}

Adds a new root for the traversal.

revwalk_repository(walk)

revwalk_repository(revwalk()) :: {:ok, repo()} | {:error, term()}

revwalk_repository(revwalk()) :: {:ok, binary()} | {:error, term()}

Returns the repository on which the given walker is operating.

revwalk_reset(walk)

revwalk_reset(revwalk()) :: revwalk()

Resets the revision walk for reuse.

revwalk_simplify_first_parent(walk)

revwalk_simplify_first_parent(revwalk()) :: :ok | {:error, term()}

Simplifies the history by first-parent.

revwalk_sorting(walk, sort_mode)

revwalk_sorting(revwalk(), [revwalk_sort()]) :: :ok | {:error, term()}

Changes the sorting mode when iterating through the repository’s contents.

revwalk_stream(walk)

revwalk_stream(revwalk()) :: {:ok, Stream.t()} | {:error, term()}

Returns a stream for the given revision walk.

signature_default(repo)

signature_default(repo()) ::
  {:ok, binary(), binary(), non_neg_integer(), non_neg_integer()}
  | {:error, term()}

Returns the default signature for the given repo.

signature_new(name, email)

signature_new(binary(), binary()) :: {:ok, binary(), binary()}

Creates a new signature with the given name and email.

signature_new(name, email, time)

signature_new(binary(), binary(), non_neg_integer()) ::
  {:ok, binary(), binary(), non_neg_integer(), non_neg_integer()}
  | {:error, term()}

Creates a new signature with the given name, email and time.

tag_author(tag)

tag_author(tag()) ::
  {:ok, binary(), binary(), non_neg_integer(), non_neg_integer()}
  | {:error, term()}

Returns the author of the given tag.

tag_list(repo)

tag_list(repo()) :: {:ok, [binary()]} | {:error, term()}

Returns all tags for the given repo.

tag_message(tag)

tag_message(tag()) :: {:ok, binary()} | {:error, term()}

Returns the message of the given tag.

tag_name(tag)

tag_name(tag()) :: {:ok, binary()} | {:error, term()}

Returns the name of the given tag.

tag_peel(tag)

tag_peel(tag()) :: {:ok, obj_type(), oid(), obj()} | {:error, term()}

Recursively peels the given tag until a non tag object is found.

tree_byid(tree, oid)

tree_byid(tree(), oid()) ::
  {:ok, integer(), atom(), binary(), binary()} | {:error, term()}

Retrieves a tree entry owned by the given tree, given its id.

tree_bypath(tree, path)

tree_bypath(tree(), Path.t()) ::
  {:ok, integer(), atom(), binary(), binary()} | {:error, term()}

Retrieves a tree entry contained in the given tree or in any of its subtrees, given its relative path.

tree_count(tree)

tree_count(tree()) :: {:ok, non_neg_integer()} | {:error, term()}

Returns the number of entries listed in the given tree.

tree_entries(tree)

tree_entries(tree()) :: {:ok, Stream.t()} | {:error, term()}

Returns all entries in the given tree.

tree_nth(tree, nth)

tree_nth(tree(), non_neg_integer()) ::
  {:ok, integer(), atom(), binary(), binary()} | {:error, term()}

Looks for a tree entry by its position in the given tree.