GitRekt.WireProtocol behaviour (git.limo v0.3.8)

Conveniences for Git transport protocol and server side commands.

This module implements version 2 of Git's wire protocol.

It functions as a very basic finite-state machine by processing incoming client requests and forwarding them to the underlying service implementation (respectively receive-pack and upload-pack).

The state machine is initialized by calling new/2 with the Git repository and command to execute. By passing incoming data to next/2, the underlying service transit to the next state. Once the client and the server are done with exchanging Git objects, the service will reach the :done state.

When processing a entire (not chunked), one can use run/2 to execute all the steps in a single call.

Callbacks

next(struct, list)

Specs

next(struct(), [term()]) :: {struct(), [term()]}

Callback used to transist a service to the next step.

skip(struct)

Specs

skip(struct()) :: struct()

Callback used to transist a service to the next step without performing any action.

Functions

decode(pkt)

Specs

decode(binary()) :: Enumerable.t()

Returns a stream of decoded PKT-LINEs for the given pkt.

done?(service)

Specs

done?(struct()) :: boolean()

Returns true if service is done; elsewise returns false.

encode(lines)

Specs

encode(Enumerable.t()) :: iolist()

Returns an PKT-LINE encoded representation of the given lines.

new(agent, executable, init_values \\ [])

Specs

new(GitRekt.GitAgent.agent(), binary(), keyword()) :: struct()

Returns a new service object for the given repo and executable.

next(service, data \\ :discovery)

Specs

next(struct(), binary() | :discovery) :: {:cont | :halt, struct(), iolist()}

Runs the given service to the next step.

pkt_line(data \\ :flush)

Specs

pkt_line(binary() | :flush) :: binary()

Returns the given data formatted as PKT-LINE

reference_discovery(agent, service, extra_capabilities \\ [])

Specs

reference_discovery(GitRekt.GitAgent.agent(), binary(), [binary()]) :: iolist()

Returns a stream describing each ref and it current value.

run(service, data \\ :discovery, opts \\ [])

Specs

run(struct(), binary() | :discovery, keyword()) :: {struct(), iolist()}

Runs all the steps of the given service at once.

skip(service)

Specs

skip(struct()) :: struct()

Sets the given service to the next logical step without performing any action.